### 2024-03-16 #6498 @editorialbot
[PRE REVIEW]: D3mirt: Descriptive Three-Dimensional Multidimensional Item Response Theory   for R
**Submitting author:** @ForsbergPyschometrics (<a href="http://orcid.org/0000-0002-5228-9729">Erik Forsberg</a>)
**Repository:** https://github.com/ForsbergPyschometrics/D3mirt
**Branch with paper.md** (empty if default branch):
**Version:** 1.1.0
