### 2024-03-16 #6498 @ForsbergPyschometrics
paper.md file has been added to the GitHub directory.

Regards,

Erik Forsberg
Psychometrician

Phone: +46(0)736007084
https://github.com/ForsbergPyschometrics

Please, consider the environment before printing this E-mail!

Den lör 16 mars 2024 kl 16:00 skrev The Open Journals editorial robot <
***@***.***>:

> ⚠️ An error happened when generating the pdf
> <https://github.com/openjournals/joss-papers/actions/runs/8308573308?check_suite_focus=true>.
> Paper file not found.
>
