### 2024-03-17 #6498 @osorensen
@ForsbergPyschometrics, as you can see, I have made some pull requests with suggested changes to the paper. Great if you can fix these before we start the review.
