### 2024-03-17 #6498 @ForsbergPyschometrics
Hi Øystein,

I have approved all the suggestions. Let me know if I can do anything more
to help.

Regards,

Erik Forsberg
Psychometrician

Phone: +46(0)736007084
https://github.com/ForsbergPyschometrics

Please, consider the environment before printing this E-mail!

Den sön 17 mars 2024 kl 11:54 skrev Øystein Sørensen <
***@***.***>:

> @ForsbergPyschometrics <https://github.com/ForsbergPyschometrics>, as you
> can see, I have made some pull requests with suggested changes to the
> paper. Great if you can fix these before we start the review.
>
