### 2024-03-17 #6498 @ForsbergPyschometrics
Done.
Regards,

Erik Forsberg
Psychometrician

Phone: +46(0)736007084
https://github.com/ForsbergPyschometrics

Please, consider the environment before printing this E-mail!

Den sön 17 mars 2024 kl 17:38 skrev Øystein Sørensen <
***@***.***>:

> Thanks @ForsbergPyschometrics <https://github.com/ForsbergPyschometrics>.
> You also need to merge the pull requests into the main branch in order for
> the source code to be updated.
>
