### 2024-03-17 #6498 @ForsbergPyschometrics
Thank you for your quick editing!

I do not understand why the comma and dot need to be added to the formulas
1 and 4. It looks like they belong to the mathematical expression when they
do not and it looks confusing. But I guess this can be changed at a later
stage.

Otherwise, I approve of the draft.

Regards,

Erik Forsberg
Psychometrician

Phone: +46(0)736007084
https://github.com/ForsbergPyschometrics

Please, consider the environment before printing this E-mail!

Den sön 17 mars 2024 kl 20:15 skrev The Open Journals editorial robot <
***@***.***>:

> 👉📄 Download article proof
> <https://raw.githubusercontent.com/openjournals/joss-papers/joss.06498/joss.06498/10.21105.joss.06498.pdf>
> 📄 View article proof on GitHub
> <https://github.com/openjournals/joss-papers/blob/joss.06498/joss.06498/10.21105.joss.06498.pdf>
> 📄 👈
>
