### 2024-03-17 #6498 @osorensen
@ForsbergPyschometrics, the point was that the equations are part of the text, even though they are displayed. In these cases it is common style to not end sentences in front of equations but instead treat them as part of the sentence.

Here are some online references on the topic:

https://www.e-education.psu.edu/styleforstudents/c4_p3.html

https://www.letpub.com/author_education_formatting_and_editing_math_equations

https://web.cs.ucdavis.edu/~amenta/w10/writingman.pdf
