### 2024-03-17 #6498 @ForsbergPyschometrics
Thank you, I understand. Well, my point was just my own taste. I trust you
on this matter.

Regards,

Erik Forsberg
Psychometrician

Phone: +46(0)736007084
https://github.com/ForsbergPyschometrics

Please, consider the environment before printing this E-mail!

Den sön 17 mars 2024 kl 20:33 skrev Øystein Sørensen <
***@***.***>:

> @ForsbergPyschometrics <https://github.com/ForsbergPyschometrics>, the
> point was that the equations are part of the text, even though they are
> displayed. In these cases it is common style to not end sentences in front
> of equations but instead treat them as part of the sentence.
>
> Here are some online references on the topic:
>
> https://www.e-education.psu.edu/styleforstudents/c4_p3.html
>
>
> https://www.letpub.com/author_education_formatting_and_editing_math_equations
>
> https://web.cs.ucdavis.edu/~amenta/w10/writingman.pdf
>
