### 2024-03-19 #6498 @JeffreyCHoover
Hi @osorensen, I'd be willing to review this submission for JOSS if you're still looking for a reviewer
