### 2024-03-21 #6498 @osorensen
@mhu48, @hendr1km, @wesleyburr, @amarkos, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
