### 2024-03-21 #6498 @amarkos
Thank you for the invitation! I'm happy to review the submission for JOSS.
