### 2024-03-22 #6523 @editorialbot
[REVIEW]: D3mirt: Descriptive Three-Dimensional Multidimensional Item Response Theory   for R
**Submitting author:** @ForsbergPyschometrics (<a href="http://orcid.org/0000-0002-5228-9729">Erik Forsberg</a>)
**Repository:** https://github.com/ForsbergPyschometrics/D3mirt
**Branch with paper.md** (empty if default branch):
**Version:** v2.0.4
**Archive:** 10.17605/OSF.IO/ZEFGW
