### 2024-03-23 #6498 @mhu48
I am kinda swamped right now. Sorry I could not find time to review, but
thanks for the invitation.

On Sat, Mar 23, 2024 at 5:12 AM The Open Journals editorial robot <
***@***.***> wrote:

> Checklists can only be created once the review has started in the review
> issue
>
