### 2024-03-28 ForsbergPyschometrics/D3mirt#12 @JeffreyCHoover
Reproducibility
In reviewing your paper, I'm unable to test the reproducibility of the results. Specifically, I'm getting an error message with the line:

`mod1 <- mirt::mirt(x, spec, itemtype = 'graded', SE = TRUE, method = 'QMCEM')`

with an error message of:

`Error: Keywords and latent trait/class names must differ from colnames(data)`

Addressing this issue will help with reproducing the results shared in your paper, evaluating the functionality of the R package, and having function examples that avoid this error message (`D3mirt()` and `plot.D3mirt()`).
