### 2024-03-28 ForsbergPyschometrics/D3mirt#13 @JeffreyCHoover
Software paper edits
Your paper read well. The one piece of feedback that I would offer is that it appears your software paper would be strengthened by including a comparison to other software/packages performing similar functions and describing how your package provides functionality above and beyond those other softwares/packages.

As a minor point, I think the in-text citation for Forsberg & Sjöberg needs to be updated. It currently reads as just "Forsberg (2024)".
