### 2024-03-28 ForsbergPyschometrics/D3mirt#14 @JeffreyCHoover
Documentation
A couple minor suggestions for improvements to the package documentation:

- including unit tests to verify the performance of the package's functions
- including a note and guidelines in the README for those wishing to contribute to the package
- reiterating the statement of need from the software paper in the package README
- ensuring that the examples in the function documentation are stand-alone; for instance, the example for `summary.modid()` begins with `g <- modid(x)` but users would have to either run the other examples first to have the correct object for `x` or they have to reference other parts of the package to correctly specify `x`
