### 2024-03-28 #6523 @JeffreyCHoover
I added a few issues in the target repo.

[#12](https://github.com/ForsbergPyschometrics/D3mirt/issues/12): Reproducibility
[#13](https://github.com/ForsbergPyschometrics/D3mirt/issues/13): Software paper edits
[#14](https://github.com/ForsbergPyschometrics/D3mirt/issues/14): Documentation
