### 2024-04-11 #6523 @ForsbergPyschometrics
Hi,

Thanks for the comments!

Paper
I certainly can add something on other software options. However, the
package has no real equivalent. There are only applications that can do
some of these things. Just so you know.
The reference is of course fixable.

Package
I am unable to reproduce the mirt-error when using the code. However, I am
planning to update the D3mirt function so that the mirt function call is
imbedded in the D3mirt code. This may fix this issue and it also means that
the external specification of the model will be unnecessary. Hence, less
risk of user mistakes. I have also created a type of test that can be used
to check the results and the graphical output that will be included in the
update.
I will, however, wait until all reviewer's are finished so that I can
address all suggestions for improvements at once.

Regards,

Erik Forsberg
Psychometrician

Phone: +46(0)736007084
https://github.com/ForsbergPyschometrics

Please, consider the environment before printing this E-mail!

Den lör 30 mars 2024 kl 12:04 skrev Øystein Sørensen <
***@***.***>:

> Thanks @JeffreyCHoover <https://github.com/JeffreyCHoover>!
> @ForsbergPyschometrics <https://github.com/ForsbergPyschometrics>, you're
> welcome to start addressing these issues right away, even though the other
> reviewers haven't started yet.
>
