### 2024-04-15 ForsbergPyschometrics/D3mirt#12 @JeffreyCHoover
Since you were unable to reproduce the error message, I tried the code again, and it worked this time. I'm closing this out since it's reproducible.
