### 2024-04-15 #6523 @JeffreyCHoover
> Hi, Thanks for the comments! Paper I certainly can add something on other software options. However, the package has no real equivalent. There are only applications that can do some of these things. Just so you know.

Ah, that makes sense. I think even a simple statement to that effect conveys the importance of the package and its contribution.

> I am unable to reproduce the mirt-error when using the code. However, I am planning to update the D3mirt function so that the mirt function call is imbedded in the D3mirt code. This may fix this issue and it also means that the external specification of the model will be unnecessary. Hence, less risk of user mistakes. I have also created a type of test that can be used to check the results and the graphical output that will be included in the update.

Since you were unable to reproduce the error message, I tried the code again, and it worked this time...I closed out that issue, since I can reproduce the examples.

> I will, however, wait until all reviewer's are finished so that I can address all suggestions for improvements at once.

Sounds good! Thanks for considering my comments!
