### 2024-05-05 ForsbergPyschometrics/D3mirt#15 @hendr1km
Equation Refs 
On Page 4 of your submission: Repeated "Equation" caused by writing Equation in front of the autoref (e.g. line 83.)
