### 2024-05-05 ForsbergPyschometrics/D3mirt#16 @hendr1km
State of the field: Do the authors describe how this software compares to other commonly-used packages?
Not described how this software compares to other commonly-used packages (eg mirt)
