### 2024-05-05 ForsbergPyschometrics/D3mirt#17 @hendr1km
No clear guidelines for third parties wishing to contribute to the software
No clear guidelines for third parties wishing to contribute to the software e.g. Using Pull requests
