### 2024-05-05 #6523 @hendr1km
[#15](https://github.com/ForsbergPyschometrics/D3mirt/issues/15):  Equation Refs in paper
[#16](https://github.com/ForsbergPyschometrics/D3mirt/issues/16):  State of field in paper
[#17](https://github.com/ForsbergPyschometrics/D3mirt/issues/17):  No clear guidelines for third parties wishing to contribute to the software in documentation (only regarding bugs and questions)
