### 2024-05-05 ForsbergPyschometrics/D3mirt#19 @smasongarrison
Functionality
JOSS review criterion: Functionality (https://github.com/openjournals/joss-reviews/issues/6523)

goodpractice picked up some issues that may impact package functionality. The largest of which is the lack of test coverage, but there are a couple of others, like avoiding sapply.  Interestingly, two of the tests couldn't even be run (cyclocomp, rcmdcheck) within goodpratice. Running those on their own, indicated that plot.D3mirt was *extremely* complex (128).

- [x] Could you add unit tests to cover basic functionalities of all functions in the package?
- [x] Please consider replacing sapply() instances with vapply(), as well as the other functions flagged in good practice.
- [x] Is it feasible to refactor the plot.D3mirt function to reduce its complexity significantly, ideally to double digits? I've had my own experiences with reducing the complexity of functions (That was from 78 to functions in the mid 30s https://github.com/R-Computing-Lab/BGmisc/issues/13 )  Simplifying this function could make the package easier to maintain and extend.

> goodpractice::gp()
Preparing: covr
Preparing: cyclocomp
Preparing: description
Preparing: lintr
  |========================================================================================================| 100%
Preparing: namespace
Preparing: rcmdcheck
── GP D3mirt ─────────────────────────────────────────────────────────────────────────────────────────────────────

It is good practice to

  ✖ write unit tests for all functions, and all package code in general. 0% of code lines are
    covered by test cases.

    R/D3mirt.R:92:NA
    R/D3mirt.R:93:NA
    R/D3mirt.R:94:NA
    R/D3mirt.R:95:NA
    R/D3mirt.R:97:NA
    ... and 546 more lines

  ✖ not use "Depends" in DESCRIPTION, as it can cause name clashes, and poor interaction with
    other packages. Use "Imports" instead.
  ✖ avoid long code lines, it is bad for readability. Also, many people prefer editor windows
    that are about 80 characters wide. Try make your lines shorter than 80 characters

    R\D3mirt.R:3:81
    R\D3mirt.R:5:81
    R\D3mirt.R:6:81
    R\D3mirt.R:7:81
    R\D3mirt.R:11:81
    ... and 281 more lines

  ✖ avoid sapply(), it is not type safe. It might return a vector, or a list, depending on
    the input data. Consider using vapply() instead.

    R\D3mirt.R:132:58
    R\D3mirt.R:133:58
    R\D3mirt.R:140:13
    R\plotD3mirt.R:334:13
    R\plotD3mirt.R:341:11
    ... and 26 more lines

  ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...), 1:NROW(...) and 1:NCOL(...) expressions.
    They are error prone and result 1:0 if the expression on the right hand side is zero. Use
    seq_len() or seq_along() instead.

    R\modid.R:104:12

  ✖ not import packages as a whole, as this can cause name clashes between the imported
    packages. Instead, import only the specific functions you need.
──────────────────────────────────────────────────────────────────────────────────────────────────────────────────
Warning messages:
1: In MYPREPS[[prep]](state, quiet = quiet) :
  Prep step for cyclomatic complexity failed.
2: In MYPREPS[[prep]](state, quiet = quiet) :
  Prep step for rcmdcheck failed.

> library(cyclocomp)
> cyclocomp_package_dir(path = ".")
── R CMD build ───────────────────────────────────────────────────────────────────────────────────────────────────
✔  checking for file 'C:\Users\smaso\AppData\Local\Temp\RtmpmqXXgl\remotes46682e3c5642\D3mirt/DESCRIPTION' ...
─  preparing 'D3mirt': (620ms)
✔  checking vignette meta-information
─  building 'D3mirt_1.1.0.tar.gz'

* installing *source* package 'D3mirt' ...
** using staged installation
** R
** data
*** moving datasets to lazyload DB
** inst
** byte-compile and prepare package for lazy loading
** help
*** installing help indices
*** copying figures
** building package indices
** installing vignettes
** testing if installed package can be loaded from temporary location
** testing if installed package can be loaded from final location
** testing if installed package keeps a record of temporary installation path
* DONE (D3mirt)
            name cyclocomp
3    plot.D3mirt       128
1         D3mirt        44
2          modid        13
7  summary.modid         7
5    print.modid         5
6 summary.D3mirt         5
4   print.D3mirt         4
