### 2024-05-05 #6523 @smasongarrison
I've finished my review as well, see https://github.com/openjournals/joss-reviews/issues/6523#issuecomment-2093465632

# Comments:
- I've created an issue for functionality improvements and echo the other reviewers suggestions (regarding the need for Community guidelines and state of the field).
- In general with the paper, the section on Multidimensional item parameters is interesting, but I think that it can be shortened considerably (especially since that information is in https://www.tandfonline.com/doi/full/10.1080/15366367.2023.2282418 ). Instead, I think that more of the paper should be dedicated to the use of similar software, even if it doesn't use the same models. Other IRT software does exist, including `mirt` and this old list to give you a sense of how much there is (last updated in 2009; https://www.umass.edu/remp/software/CEA-652.ZH-IRTSoftware.pdf)
