### 2024-05-06 #6523 @ForsbergPyschometrics
Ok, I will do that. Thank you for your comments!
Regards,

Erik Forsberg
Psychometrician

Phone: +46(0)736007084
https://github.com/ForsbergPyschometrics

Please, consider the environment before printing this E-mail!

Den mån 6 maj 2024 kl 14:06 skrev Øystein Sørensen ***@***.***
>:

> @ForsbergPyschometrics <https://github.com/ForsbergPyschometrics>, I
> suggest you start addressing the issue now, as we have three reviews
> already.
>
