+++ data-raw/angles.R
@@ -1,0 +1,80 @@
+## code to prepare `angles` dataset goes here
+a <- matrix(c(1, 0, 0,
+              0, 1, 0,
+              0, 0, 1,
+              .5, .5, .5), ncol = 4)
+
+b <- matrix(c(-1, 0, 0,
+              0, -1, 0,
+              0, 0, -1,
+              .5, .5, .5), ncol = 4)
+
+c <- matrix(c(1/2, sqrt(2)/2, sqrt(3)/2, # standardvinklar mot x
+              sqrt(3)/2, sqrt(2)/2, 1/2,
+              0, 0, 0,
+              .5, .5, .5), ncol = 4)
+
+d <- matrix(c(-1/2, -sqrt(2)/2, -sqrt(3)/2, # standardvinklar mot -x
+              sqrt(3)/2, sqrt(2)/2, 1/2,
+              0, 0, 0,
+              .5, .5, .5), ncol = 4)
+
+e <- matrix(c(0, 0, 0, # standardvinklar mot z
+              1/2, sqrt(2)/2, sqrt(3)/2,
+              sqrt(3)/2, sqrt(2)/2, 1/2,
+              .5, .5, .5), ncol = 4)
+
+f <- matrix(c(0, 0, 0, # standardvinklar mot -z
+              1/2, sqrt(2)/2, sqrt(3)/2,
+              -sqrt(3)/2, -sqrt(2)/2, -1/2,
+              .5, .5, .5), ncol = 4)
+
+g <- matrix(c(1/2, sqrt(2)/2, sqrt(3)/2, # standardvinklar mot x, -y
+              -sqrt(3)/2, -sqrt(2)/2, -1/2,
+              0, 0, 0,
+              .5, .5, .5), ncol = 4)
+
+h <- matrix(c(-1/2, -sqrt(2)/2, -sqrt(3)/2, # standardvinklar mot -x, -y
+              -sqrt(3)/2, -sqrt(2)/2, -1/2,
+              0, 0, 0,
+              .5, .5, .5), ncol = 4)
+
+i <- matrix(c(0, 0, 0, # standardvinklar mot z, -y
+              -1/2, -sqrt(2)/2, -sqrt(3)/2,
+              sqrt(3)/2, sqrt(2)/2, 1/2,
+              .5, .5, .5), ncol = 4)
+
+
+j <- matrix(c(0, 0, 0, # standardvinklar mot -z, -y
+              -1/2, -sqrt(2)/2, -sqrt(3)/2,
+              -sqrt(3)/2, -sqrt(2)/2, -1/2,
+              .5, .5, .5), ncol = 4)
+
+k <- matrix(c(sqrt(1/3), -sqrt(1/3), -sqrt(1/3), sqrt(1/3), # .57 positiv
+              sqrt(1/3), sqrt(1/3), sqrt(1/3), sqrt(1/3),
+              sqrt(1/3), sqrt(1/3), -sqrt(1/3), -sqrt(1/3),
+              .5, .5, .5, .5), ncol = 4)
+
+l <- matrix(c(sqrt(1/3), -sqrt(1/3), -sqrt(1/3), sqrt(1/3), # .57 negativ
+              -sqrt(1/3), -sqrt(1/3), -sqrt(1/3), -sqrt(1/3),
+              sqrt(1/3), sqrt(1/3), -sqrt(1/3), -sqrt(1/3),
+              .5, .5, .5, .5), ncol = 4)
+
+m <- matrix(c(sqrt(2)/2, -sqrt(2)/2, -sqrt(2)/2, sqrt(2)/2, # 45 i planet alla riktningar
+              0, 0, 0, 0,
+              sqrt(2)/2, sqrt(2)/2, -sqrt(2)/2, -sqrt(2)/2,
[…]
