+++ R/D3mirt.R
@@ -1,32 +1,54 @@
 #' 3D DMIRT Model Estimation
 #'
-#' @description Descriptive multidimensional item response theory model estimation (DMIRT; Reckase, 2009, 1985, Reckase & McKinley, 1991) for dichotomous and polytomous items restricted to three dimensions.
+#' @description Descriptive multidimensional item response theory model estimation (DMIRT; Reckase, 2009, 1985, Reckase and McKinley, 1991) for dichotomous and polytomous items restricted to three dimensions.
 #'
 #' @param x A data frame with rows for items and columns for model parameters or an S4 object of class 'SingleGroupClass' exported from [mirt::mirt] (Chalmers, 2012).
 #' Regarding the data frame, the number of columns must be more than or equal to 4, i.e., three columns with (\emph{a}) parameters and at least one column for difficulty (\emph{d}) parameters.
-#' @param constructs Optional. Nested lists with integer values indicating construct. The default is `constructs = NULL`.
+#' @param modid Argument that either takes in the two model identification items as a combined vector. The user can also use nested list consisting of item indicators to fit an orthogonal model.
+#' The default is `modid = NULL`.
+#' @param model The user has the option of imputing a model specification schema that is used in the call to [mirt::mirt] (Chalmers, 2012).
+#' The default is `model = NULL`.
+#' @param con.items Optional. Nested lists with integer values as item indicators to identify constructs. The default is `con.items = NULL`.
+#' @param con.sph Optional. Nested lists of spherical angles to identify constructs. The default is `con.sph = NULL`.
+#' @param itemtype What item type to use in the function call. Available options are `"2PL"` and `"graded"`. The default is `itemtype = "graded"`.
+#' @param method.mirt Estimation algorithm for [mirt::mirt] (Chalmers, 2012) for fitting the model. The default is `method.mirt = "QMCEM"`.
+#' @param method.fscores Factor estimation algorithm for [mirt::fscores] (Chalmers, 2012) for extracting respondent trait scores. The default is `method.fscores = "EAP"`.
+#' @param QMC Integration method for [mirt::fscores] (Chalmers, 2012). The default is `QMC = TRUE`.
 #'
 #' @importFrom mirt mirt
 #'
 #' @details The `D3mirt()` function takes in model parameters from a compensatory three-dimensional multidimensional two-parameter logistic model (M2PL) or a multidimensional graded
-#' response model (MGRM), either in the form of a data frame or an S4 object of class 'SingleGroupClass' exported from [mirt::mirt] (Chalmers, 2012) function fitted in accordance with the descriptive item response theory model specifications described below.
+#' response model (MGRM), either in the form of a data frame with item data, or a data frame with factor loadings or an S4 object of class 'SingleGroupClass' exported from [mirt::mirt] (Chalmers, 2012) function fitted in accordance with descriptive item response theory model specifications (see package vignette).
 #' The function returns DMIRT estimates that can be visualized with [D3mirt::plot] that graph vector arrows representing item response characteristics in a three-dimensional space.
 #' Regarding the former, this includes visualization of the single multidimensional discrimination (MDISC) parameter and the multidimensional difficulty (MDIFF) parameters (Reckase, 2009, 1985; Reckase & McKinley, 1991).
[…]
