+++ R/summarymodid.R
@@ -12,13 +12,13 @@
 #' # Identify the DMIRT model using a three-factor EFA with modid()
 #' # x can be a data frame with item data or item factor loadings.
 #' # In the case of the latter, set argument 'efa' to 'FALSE'
-#' g <- modid(x)
+#' id <- modid(x)
 #'
 #' # Call to summary
-#' summary(x)
+#' summary(id)
 #'
 #' # Call to summary rounded off to 2 digits
-#' summary(x, digits = 2)
+#' summary(id, digits = 2)
 #' }
 #' @export
 summary.modid <- function(object, ..., digits = 4){
