+++ data-raw/angles.R
@@ -66,10 +66,12 @@ m <- matrix(c(sqrt(2)/2, -sqrt(2)/2, -sqrt(2)/2, sqrt(2)/2, # 45 i planet alla r
               .5, .5, .5, .5), ncol = 4)
 
 s <- data.frame(Theta = c(0, NaN, 90, 180, NaN, -90, 0, 0, 0, 180, 180, 180, 90, 90, 90,
-                          -90, -90, -90, 0, 0, 0, 180, 180, 180, 90, 90, 90, -90, -90, -90, 45, 135, -135, -45, 45, 135, -135, -45, 45, 135, -135, -45),
-                Phi = c(90, 0,  90,  90, 180,  90,  30,  45,  60,  30,  45,  60, 60, 45, 30, 60, 45, 30, 150, 135, 120, 150, 135, 120, 120, 135, 150, 120,
-                        135, 150,  54.73561,  54.73561,  54.73561,  54.73561, 125.26439, 125.26439,
-                        125.26439, 125.26439,  90,  90,  90, 90))
+                          -90, -90, -90, 0, 0, 0, 180, 180, 180, 90, 90, 90, -90, -90,
+                          -90, 45, 135, -135, -45, 45, 135, -135, -45, 45, 135, -135, -45),
+                Phi = c(90, 0,  90,  90, 180,  90,  30,  45,  60,  30,  45,  60, 60, 45,
+                        30, 60, 45, 30, 150, 135, 120, 150, 135, 120, 120, 135, 150, 120,
+                        135, 150,  54.73561,  54.73561,  54.73561,  54.73561, 125.26439,
+                        125.26439, 125.26439, 125.26439,  90,  90,  90, 90))
 
 z <- rbind(a, b, c, d, e, f, g, h, i, j, k, l, m)
 z <- cbind(z, s)
