+++ R/printD3mirt.R
@@ -22,18 +22,25 @@ print.D3mirt <- function(x, ...){
   tab1 <- as.data.frame(x$loadings)
   tab2 <- as.data.frame(x$diff)
   tab1 <- as.data.frame(cbind(tab1, tab2))
-  if (!is.null(x$c.dir.cos)){
-    cat(paste("\nD3mirt:", nrow(tab1), "items and", ncol(tab2), "levels of difficulty\n\n"))
-    cat(paste("Constructs:\n"))
-    for (i in seq_along(x$c)){
-      n <- unlist(x$c[i])
-      z <- sapply(n, function (x){
-        q <- as.character(rownames(tab1[x,]))
-      })
-      cat(paste("Vector ", i, ": ", paste(z, collapse=", "), "\n", sep = ""))
+  if (length(x$diff) > 1){
+    cat(paste("\nD3mirt:", nrow(tab1), "items and", length(tab2), "levels of difficulty\n\n"))
+    } else {
+    cat(paste("\nD3mirt:", nrow(tab1), "items and", length(tab2), "level of difficulty\n\n"))
     }
-    cat(paste("\n"))
-  } else {
-    cat(paste("\nD3mirt:", nrow(tab1), "items and", ncol(tab2), "levels of difficulty\n\n"))
+    if (!is.null(x$con.items)){
+    cat(paste("Constructs:\n"))
+    for (i in seq_along(x$con.items)){
+      n <- unlist(x$con.items[i])
+      z <- as.character(rownames(tab1[n, ]))
+        cat(paste("Item vector ", i, ": ", paste(z, collapse=", ", sep = ""), "\n", sep = ""))
   }
+  }
+   if (!is.null(x$con.sph)){
+      cat(paste("Constructs:\n"))
+      for (i in seq_along(x$con.sph)){
+        n <- unlist(x$con.sph[i])
+          cat(paste("Spherical coordinate vector ", i, ": ", paste(n[1], "°", ", ", collapse="", sep = ""), paste(n[2], "°", collapse="", sep = ""), "\n", sep = ""))
+      }
+  }
+  cat(paste("\n"))
 }

[…]
