+++ R/summaryD3mirt.R
@@ -68,7 +68,7 @@ summary.D3mirt <- function(object, ..., digits = 4){
     cat(paste("Constructs\n"))
       for (i in seq_along(object$con.sph)){
         n <- unlist(object$con.sph[i])
-        cat(paste("Spherical coordinate vector ", i, ": ", paste(n[1], "°", ", ", collapse="", sep = ""), paste(n[2], "°", collapse="", sep = ""), "\n", sep = ""))
+        cat(paste("Spherical coordinate vector ", i, ": ", paste(n[1], ", ", collapse="", sep = ""), paste(n[2], collapse="", sep = ""), "\n", sep = ""))
       }
     }
     cat(paste("\n"))

[…]
