+++ R/plotD3mirt.R
@@ -31,7 +31,6 @@
 #' Note, a symmetric model can be created easily by adding a single numeric in the `axis.length` argument (e.g., `axis.length = 4`) because the function repeats the last value in the vector to cover all axis points.
 #' The default is `axis.length = NULL`.
 #' @param axis.points Color of axis points for the `points3d()` function. The default is `axis.points = "black"`.
-#' @param points Logical, if axis from `points3d()` should have end points. The default is `points = TRUE`.
 #' @param axis.ticks Logical, if axis ticks from the `axis3d()` function should be plotted. The default is `axis.ticks = TRUE`.
 #' @param nticks Number of ticks for `axis3d()`.
 #' The function repeats the last numeric value in the vector to cover all axis.
@@ -212,7 +211,7 @@ plot.D3mirt <- function (x, scale = FALSE, hide = FALSE, ind.scores = FALSE, dif
                          constructs = FALSE, construct.lab = NULL, adjust.lab = c(0.5, -0.8),
                          x.lab = "X", y.lab="Y", z.lab="Z", title="", line = -5, font = 1, cex = 1, font.col = "black",
                          axis.scalar = 1.1, axis.length = NULL, axis.points = "black",
-                         points = TRUE, axis.ticks = TRUE, nticks = 4,  width.rgl.x = 1040, width.rgl.y= 1040, view = c(15, 20, 0.6),
+                         axis.ticks = TRUE, nticks = 4,  width.rgl.x = 1040, width.rgl.y= 1040, view = c(15, 20, 0.6),
                          show.plane = TRUE, plane.col = "grey80", background = "white",
                          type = "rotation", col = c("black", "grey20", "grey40", "grey60", "grey80"),
                          arrow.width = 0.6, n = 20, theta = 0.2, barblen = 0.03,
@@ -278,11 +277,9 @@ plot.D3mirt <- function (x, scale = FALSE, hide = FALSE, ind.scores = FALSE, dif
     rgl::axis3d('y', pos = c(0, 0, 0), ticks = TRUE, nticks=nticks[2], cex = cex, font = font, color = font.col)
     rgl::axis3d('z',pos = c(0, 0, 0), ticks = TRUE, nticks=nticks[3], cex = cex, font = font, color = font.col)
   }
-  if (points == TRUE){
-    axes <- rbind(c(xaxis[2], 0, 0), c(0, yaxis[2], 0),
+  axes <- rbind(c(xaxis[2], 0, 0), c(0, yaxis[2], 0),
                   c(0, 0, zaxis[2]))
-    rgl::points3d(axes, color = axis.points, cex = cex)
-  }
+  rgl::points3d(axes, color = axis.points, cex = cex)
   rgl::text3d(axes, text = c(x.lab, y.lab, z.lab), color = font.col,
               adj = c(0.5, -0.8), font = font, cex = cex)
   rgl::title3d(main= title,line= line, cex = cex, font = font, color = font.col)
@@ -295,6 +292,10 @@ plot.D3mirt <- function (x, scale = FALSE, hide = FALSE, ind.scores = FALSE, dif
   if (hide == FALSE){
     if (scale == FALSE){
       vec <- x$dir.vec
+    }
+    if (scale == TRUE) {
+      vec <- x$scal.vec
+    }
       if (!is.null(items)){
         if(any(!items <= nrow(x$loadings))) stop("The items argument contains one or more item indicators higher than the total number of items")
         if (any(duplicated(items))) stop("The items argument has duplicate elements")
@@ -444,105 +445,6 @@ plot.D3mirt <- function (x, scale = FALSE, hide = FALSE, ind.scores = FALSE, dif
           }
         }
       }
-    } else {
-      vec <- x$scal.vec
-      if (!is.null(items)){
[…]
