+++ tests/testthat/test-D3mirt.R
@@ -1,7 +1,7 @@
-test_that("Test unit D3mirt", {
+test_that("Test unit D3mirt and plot", {
   data(anes0809offwaves)
-  anes <- anes0809offwaves[,3:22]
-  x <- D3mirt(anes, modid= c("W7Q3", "W7Q20"), con.items = list(c(1:10), (11:15), c(16:20)))
+  anes1 <- as.matrix(anes0809offwaves[,3:22])
+  x <- D3mirt(anes1, modid= c("W7Q3", "W7Q20"), con.items = list(c(1:10), (11:15), c(16:20)))
   expect_s3_class(x, "D3mirt")
   expect_snapshot(x)
   expect_snapshot_output(x)
@@ -12,23 +12,51 @@ test_that("Test unit D3mirt", {
   for (i in seq_along(row)){
     expect_equal(row[i,1], 1)
   }
-  plot(x, constructs = TRUE)
+  anes2 <- data.frame(ifelse(anes > 4 ,1 , 0))
+  y <- D3mirt(anes2, modid= c("W7Q3", "W7Q20"))
+  row <- as.matrix(y$dir.cos)
+  row <- rowSums(row^2)
+  row <- matrix(row)
+  for (i in seq_along(row)){
+    expect_equal(row[i,1], 1)
+  }
+  plot(x, constructs = TRUE, ind.scores = TRUE, title = "Plot Test 1")
   p <- rgl::scene3d()
   expect_snapshot(p)
-  anes2 <- anes0809offwaves
-  z <- data.frame(cbind(x$fscores, anes2[, 1]))
+  plot(y, title = "Plot Test 2")
+  p <- rgl::scene3d()
+  expect_snapshot(p)
+  plot(y, items = c(1:10), title = "Plot Test 3")
+  p <- rgl::scene3d()
+  expect_snapshot(p)
+  plot(y, items = c(1:10), item.lab = c(1:10), title = "Plot Test 4")
+  p <- rgl::scene3d()
+  expect_snapshot(p)
+  plot(x, diff.level = 3, title = "Plot Test 5")
+  p <- rgl::scene3d()
+  expect_snapshot(p)
+  plot(x, items = c(15:20), diff.level = 3, title = "Plot Test 6")
+  p <- rgl::scene3d()
+  expect_snapshot(p)
+  plot(x, items = c(15:20), diff.level = 3, item.lab = c(1:6), title = "Plot Test 7", axis.scalar = 2.1, show.plane = FALSE)
+  p <- rgl::scene3d()
+  expect_snapshot(p)
+  anes3 <- anes0809offwaves
+  z <- data.frame(cbind(x$fscores, anes3[, 1]))
   z1 <- subset(z, z[, 4] <= 30)
-  plot(x, constructs = TRUE, construct.lab = c(1:3), adjust.lab = c(0.6, -0.9), font = 2, item.lab = c(1:20),
+  plot(x, constructs = TRUE, scale = TRUE, construct.lab = c("Con 1", "Con 2", "Con 3"), adjust.lab = c(0.6, -0.9), font = 2, item.lab = c(1:20),
        line = -4, width.rgl.x = 1500, width.rgl.y = 1500, view = c(45, 45, 0.5),
-       cex = 2, axis.length = 4, x.lab = "X", y.lab="Y", z.lab="Z", title = "Test Unit", font.col = "pink", axis.points = "green",
+       cex = 2, axis.length = 4, x.lab = "A", y.lab="B", z.lab="C", title = "Plot Test 8", font.col = "pink", axis.points = "green",
        nticks = 8, plane.col = "purple", background = "black", col = c("red", "purple", "cyan", "green", "yellow"),
[…]
