### 2024-06-12 #6523 @ForsbergPyschometrics
Hi!
Thanks for reaching out.

I have been very busy but I started working on the application and the
submission to your journal full time last week. I think I'm getting ready
to resubmit, but I was a bit unsure about the " goodpractice" function and
the unit test for all functions. I have, amongst other things, used
"expect_snapshot" from the "usethis" package for R to check the functions,
and it seems as if goodpractice does not fully understand what
expact_snapshot does. But as far as I can see, the test units in D3mirt now
cover all functions (will need to be double-checked, of course), including
all function options, as well as check the central calculations, even if
good practice still says the test units cover very little. I even included
a special data frame with standard angles that is used for checking the
calculations and the graphical output.

So my question is how much do I need to cover with test units in my code?
Do I need to follow goodpractice fully?

The latest updates on D3mirt are available via the dedicated GitHub page.

Regards,

Erik Forsberg
Psychometrician

Phone: +46(0)736007084
https://github.com/ForsbergPyschometrics

Please, consider the environment before printing this E-mail!

Den ons 12 juni 2024 kl 11:34 skrev Øystein Sørensen <
***@***.***>:

> 👋 @ForsbergPyschometrics <https://github.com/ForsbergPyschometrics>, how
> is it going with the revision? Absolute fine if it takes some time, but
> feel free to reach out if you need any input.
>
