### 2024-06-12 #6523 @osorensen
@ForsbergPyschometrics, it sounds like you are doing this very well. My experience with goodpractice is that it often gives false positives, and it's really hard to make it completely satisfied. Furthermore, there is no requirement in JOSS that you use the goodpractice package. [Our guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html#tests) only say the following:

```
Tests[](https://joss.readthedocs.io/en/latest/review_criteria.html#tests)
Authors are strongly encouraged to include an automated test suite covering the core functionality of their software.

Good: An automated test suite hooked up to continuous integration (GitHub Actions, Circle CI, or similar)
OK: Documented manual steps that can be followed to objectively check the expected functionality of the software (e.g., a sample input file to assert behavior)
Bad (not acceptable): No way for you, the reviewer, to objectively assess whether the software works
```

So my interpretation is that you are "Good", but first it will be up to the reviewers to say what they think.
