+++ tests/testthat/test-D3mirt.R
@@ -4,7 +4,7 @@ test_that("Test unit D3mirt and plot", {
   x <- D3mirt(anes1, modid= c("W7Q3", "W7Q20"), con.items = list(c(1:10), (11:15), c(16:20)))
   expect_s3_class(x, "D3mirt")
   expect_snapshot(x)
-  expect_snapshot_output(x)
+  expect_snapshot(print(x))
   expect_snapshot(summary(x))
   row <- as.matrix(x$dir.cos)
   row <- rowSums(row^2)
@@ -12,35 +12,71 @@ test_that("Test unit D3mirt and plot", {
   for (i in seq_along(row)){
     expect_equal(row[i,1], 1)
   }
-  anes2 <- data.frame(ifelse(anes > 4 ,1 , 0))
-  y <- D3mirt(anes2, modid= c("W7Q3", "W7Q20"))
+  anes2 <- data.frame(ifelse(anes1 > 4 ,1 , 0))
+  y <- D3mirt(anes2, modid= c("W7Q3", "W7Q20"), itemtype = "2PL")
   row <- as.matrix(y$dir.cos)
   row <- rowSums(row^2)
   row <- matrix(row)
   for (i in seq_along(row)){
     expect_equal(row[i,1], 1)
   }
-  plot(x, constructs = TRUE, ind.scores = TRUE, title = "Plot Test 1")
+
+  plot(x, constructs = TRUE, ind.scores = TRUE, title = "Plot Test 1.1")
   p <- rgl::scene3d()
   expect_snapshot(p)
-  plot(y, title = "Plot Test 2")
+
+  plot(x, item.names = FALSE, title = "Plot Test 1.2")
   p <- rgl::scene3d()
   expect_snapshot(p)
-  plot(y, items = c(1:10), title = "Plot Test 3")
+
+  plot(y, title = "Plot Test 2.1")
   p <- rgl::scene3d()
   expect_snapshot(p)
-  plot(y, items = c(1:10), item.lab = c(1:10), title = "Plot Test 4")
+
+  plot(y, item.names = FALSE, title = "Plot Test 2.2")
   p <- rgl::scene3d()
   expect_snapshot(p)
-  plot(x, diff.level = 3, title = "Plot Test 5")
+
+  plot(y, items = c(1:10), title = "Plot Test 3.1")
   p <- rgl::scene3d()
   expect_snapshot(p)
-  plot(x, items = c(15:20), diff.level = 3, title = "Plot Test 6")
+
+  plot(y, item.names = FALSE, items = c(1:10), title = "Plot Test 3.2")
   p <- rgl::scene3d()
   expect_snapshot(p)
-  plot(x, items = c(15:20), diff.level = 3, item.lab = c(1:6), title = "Plot Test 7", axis.scalar = 2.1, show.plane = FALSE)
+
+  plot(y, items = c(1:10), item.lab = c(1:10), title = "Plot Test 4.1")
   p <- rgl::scene3d()
   expect_snapshot(p)
+
+  plot(y, item.names = FALSE, items = c(1:10), item.lab = c(1:10), title = "Plot Test 4.2")
+  p <- rgl::scene3d()
+  expect_snapshot(p)
+
+  plot(x, diff.level = 3, title = "Plot Test 5.1")
+  p <- rgl::scene3d()
+  expect_snapshot(p)
+
+  plot(x, item.names = FALSE, diff.level = 3, title = "Plot Test 5.2")
+  p <- rgl::scene3d()
+  expect_snapshot(p)
+
+  plot(x,  items = c(15:20), diff.level = 1, title = "Plot Test 6.1")
+  p <- rgl::scene3d()
+  expect_snapshot(p)
+
[…]
