### 2024-06-12 ff5b4ff
Update D3mirt.R
