### 2024-06-12 cb18225
Update test-D3mirt.R
