+++ DESCRIPTION
@@ -10,14 +10,14 @@ URL: https://github.com/ForsbergPyschometrics/D3mirt
 BugReports: https://github.com/ForsbergPyschometrics/D3mirt/issues
 Depends:
     R (>= 3.6.0),
-    rgl (>= 1.0.1)
+    rgl (>= 1.3.1)
 Imports:
     mirt
 Suggests:
     knitr,
     rmarkdown,
     stats,
-    testthat (>= 3.0.0)
+    testthat (>= 3.2.1)
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.2.3
