+++ DESCRIPTION
@@ -20,7 +20,7 @@ Suggests:
     testthat (>= 3.2.1)
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
-RoxygenNote: 7.2.3
+RoxygenNote: 7.3.1
 LazyData: true
 VignetteBuilder: knitr
 Config/testthat/edition: 3
