+++ R/plotD3mirt.R
@@ -8,7 +8,7 @@
 #' @param diff.level Optional. Plotting of a single level of difficulty indicated by an integer.
 #' @param items Optional. The user can input a list of integers indicating what item vector arrows will be visible while the remaining item vector arrows are hidden.
 #' @param item.names Logical, if item labels should be plotted. The default is `item.names = TRUE`.
-#' @param item.lab Optional. String vector of item names that will override row names extracted from the data frame. Note, row names are not overwritten.
+#' @param item.lab Optional. String vector of item names that will override row names extracted from the data frame. Note that row names are not overwritten.
 #' Instead, the string vector in `item.lab` prints item labels on the item vector arrows currently displayed following the order of item vector arrows in the graphical output.
 #' For example, when plotting in the default mode (plotting all item vectors) the labels will follow the order of the items in the data frame.
 #' If a selection of items is plotted with `items`, e.g., `items = c(24,34,25)`, then the item labels will be displayed following the order of the vector in `items` left to right.
@@ -16,7 +16,7 @@
 #' @param constructs Logical, if construct vector arrows should be plotted. The default is `constructs = FALSE`.
 #' @param construct.lab Optional. String vector of names for constructs, similar to `item.lab`.
 #' @param adjust.lab Vector of parameters for the position of the item and construct labels for the `text3d` function.
-#' The first value is for horizontal adjustment and the second is for vertical adjustment. The default is `adjust.lab = c(0.5, -0.8)`.
+#' The first value is for horizontal adjustment, and the second is for vertical adjustment. The default is `adjust.lab = c(0.5, -0.8)`.
 #' @param x.lab Labels for x-axis, the default is `x.lab = "X"`.
 #' @param y.lab Labels for y-axis, the default is `y.lab = "Y"`.
 #' @param z.lab Labels for y-axis, the default is `z.lab = "Z"`.
@@ -24,11 +24,11 @@
 #' @param line  Title placement for `title3d()`. The default is `line = -5`.
 #' @param font A numeric font number from 1 to 5, the default is `font = 1`. See [rgl::text3d] for more on font options.
 #' @param cex A numeric character expansion value to adjust font size, the default is `cex = 1`.
-#' @param font.col Color of axes, numbers and fonts. The default is `font.col = "black"`.
+#' @param font.col Color of axes, numbers, and fonts. The default is `font.col = "black"`.
 #' @param axis.scalar Scalar multiple for adjusting the length of all axes (x, y, z) in the 3D model proportionally. The default is `axis.scalar = 1.1`.
 #' @param axis.length Optional. For adjusting the length of the axis manually by entering a numeric or a numeric vector.
 #' For instance, c(3,2,4,3,3,2) indicate axis coordinates x = 3, -x = 3, y = 4, -y = 3, z = 3, -z = 2.
-#' Note, a symmetric model can be created easily by adding a single numeric in the `axis.length` argument (e.g., `axis.length = 4`) because the function repeats the last value in the vector to cover all axis points.
+#' Note that a symmetric model can be created easily by adding a single numeric in the `axis.length` argument (e.g., `axis.length = 4`) because the function repeats the last value in the vector to cover all axes points.
 #' The default is `axis.length = NULL`.
 #' @param axis.points Color of axis points for the `points3d()` function. The default is `axis.points = "black"`.
 #' @param axis.ticks Logical, if axis ticks from the `axis3d()` function should be plotted. The default is `axis.ticks = TRUE`.
@@ -49,14 +49,14 @@
 #' @param theta Opening angle of barbs for vector arrows from `arrow3d()`. The default is `theta = 0.2`.
[…]
