+++ R/modid.R
@@ -5,8 +5,8 @@
 #' @param x A data frame with item data or item factor loadings that fit the multidimensional graded response model (MGRM) or the multidimensional 2-parameter logistic model (M2PL).
 #' @param efa Logical, if the data should be explored with exploratory factor analysis (EFA). The default is `efa = TRUE`.
 #' @param factors The number of factors for the exploratory factor analysis. The default is `factors = 3`.
-#' @param lower The lower bound for the item pool, calculated using the standard deviation of scaled item factor loadings. The default is `lower = 0.5`.
-#' @param upper The upper bound for the filtering of absolute sum scores less than or equal to the indicated value. The default is `upper = .10`
+#' @param lower The lower bound for the item pool calculated using the standard deviation of scaled item factor loadings. The default is `lower = 0.5`.
+#' @param upper The upper bound for filtering absolute sum scores less than or equal to the indicated value. The default is `upper = .10`
 #' @param fac.order Optional. Users can override the automatic sorting of factors by manually indicating factor order with integer values, e.g., `c(2,1,3)` to start with the second factor (or column) in data frame x, followed by the first factor (or column) in x, and then lastly the third factor (or column).
 #' The default is `fac.order = NULL`.
 #' @param itemtype The item model for the exploratory factor analysis. Note, only item type 'graded' (for the MGRM) or '2PL' (for the M2PL) are allowed. The default is `itemtype = "graded"`. See [mirt::mirt] (Chalmers, 2012) for more on item models.
@@ -16,25 +16,25 @@
 #'
 #' @importFrom mirt mirt
 #'
-#' @details Before performing DMIRT analysis it is necessary to identify the compensatory model (Reckase, 2009).
+#' @details Before performing DMIRT analysis, it is necessary to identify the compensatory model (Reckase, 2009).
 #' For a three-dimensional model, this implies that two items must be chosen and their loadings restricted as follows.
 #' The first item is fixed not to load on the second and third axes (y and z), while the second item is fixed not to load on the third axis (z).
-#' If this can be achieved, it is possible to create a three-dimensional DMIRT model that reflects the data properly.
+#' If this can be achieved, it is possible to create a three-dimensional DMIRT model that reflects the data correctly.
 #'
 #' The `modid()` function can help by suggesting what items to use for the latter purpose.
-#' The function does this by first performing an EFA on the data and then selecting the strongest loading items, following the order of strength of the factors and in accordance with the statistical assumptions described above.
-#' This orders the entire model so that the strongest loading item, from the strongest factor, always aligns with the x-axis and the other items follow thereon.
-#' Note, the `modid()` function is not limited to three-dimensional analysis and can be used to identify a DMIRT model on any number of dimensions.
+#' The function does this by first performing an EFA on the data and then selecting the strongest loading items, following the order of strength of the factors and following the statistical assumptions described above.
+#' This orders the entire model so that the strongest loading item, from the strongest factor, always aligns with the x-axis, and the other items follow thereon.
+#' Note that the `modid()` function is not limited to three-dimensional analysis and can be used to identify a DMIRT model on any number of dimensions.
 #'
 #' Because `D3mirt` analysis is based on the M2PL and the MGRM, it is recommended to use multidimensional item response theory EFA methods, such as the EFA option in [mirt::mirt] (Chalmers, 2012) with `itemtype = 'graded'` or `'2PL'`, so that the EFA is performed with the proper item model.
 #' For this reason, the `mirt()` function is integrated into `modid()` so that the user needs only to provide the data frame containing empirical item data in the first argument in the call to the function.
-#' Accordingly, in the default mode (`efa = TRUE`), using raw item data, the function performs an EFA with three factors as default (`factors = 3`), and then finishes with the model identification.
[…]
