+++ R/summarymodid.R
@@ -9,9 +9,12 @@
 #' @author Erik Forsberg
 #' @examples
 #' \dontrun{
-#' # Identify the DMIRT model using a three-factor EFA with modid()
-#' # x can be a data frame with item data or item factor loadings.
-#' # In the case of the latter, set argument 'efa' to 'FALSE'
+#' # Load data
+#' data("anes0809offwaves")
+#' x <- anes0809offwaves
+#' x <- x[,3:22] # Remove columns for age and gender
+#'
+#' # Identify the DMIRT model
 #' id <- modid(x)
 #'
 #' # Call to summary
