+++ R/printD3mirt.R
@@ -9,13 +9,16 @@
 #'
 #' @examples
 #' \dontrun{
+#' # Load data
+#' data("anes0809offwaves")
+#' x <- anes0809offwaves
+#' x <- x[, 3:22] # Remove columns for age and gender
+#'
 #' # Call D3mirt()
-#' # The first argument can be a data frame with model parameters
-#' # or an S4 object of class 'SingleGroupClass' exported from mirt::mirt
-#' g <- D3mirt(mod1)
+#' mod <- D3mirt(x, modid = c("W7Q3", "W7Q20"))
 #'
 #' # Print model summary
-#' print(g)
+#' print(mod)
 #' }
 #' @export
 print.D3mirt <- function(x, ...){
