+++ man/plot.D3mirt.Rd
@@ -191,7 +191,7 @@ This often means that the axes of the model may need to be adjusted, which can b
 
 The user has the option of adding constructs to the graphical output with \code{constructs = TRUE} (see the documentation for \link{D3mirt} or the package vignette regarding constructs).
 Other options include plotting one level of difficulty at a time with the \code{diff.level} argument if polytomous items are used in the model.
-Item row names are displayed by default, but the user has the option of adding new item labels for the items with \code{item.lab}, as well as labeling constructs with \code{construct.lab}.
+Item names are displayed by default, but the user has the option of adding new item labels for the items with \code{item.lab}, as well as labeling constructs with \code{construct.lab}.
 
 Regarding the interpretation of results, the angle of the vector arrows indicates what traits, located along the orthogonal axes, an item can be said to describe (Reckase, 2009, 1985, Reckase & McKinley, 1991).
 For instance, an item located at 0 degrees seen from the x-axis, and 90 degrees as seen from the y and z-axis, only describes trait x.
@@ -206,7 +206,7 @@ This is comparable to trace lines turning horizontal in a unidimensional item re
 
 The plot function can also display respondent scores in the three-dimensional model space, represented as spheres whose coordinates are derived from the respondent's factor scores.
 This allows for a profile analysis in which respondent rows are separated or selected conditioned on some external criteria.
-To do this, the user must first extract respondent factor scores with \link[mirt:fscores]{mirt::fscores}(Chalmers, 2012) and then use some selection process to separate or subset respondent rows.
+To do this, the user must first extract respondent factor scores with \link[mirt:fscores]{mirt::fscores} (Chalmers, 2012) and then use some selection process to separate or subset respondent rows.
 The resulting data frame is used in the \code{profiles} argument.
 If desired, a confidence interval can be added to the spheres by setting \code{ci = TRUE}.
 A general advice is to hide vector arrows with \code{hide = TRUE} when analyzing respondent profiles to avoid visual cluttering.
@@ -227,25 +227,26 @@ con <- list(c(1,2,3,4,5,6,7,8,9,10),
             c(11,12,13,14),
             c(15,17,18,19,20))
 mod <- D3mirt(x, modid = c("W7Q3", "W7Q20"), con.items = con)
+
+# Plot RGL device with constructs invisible
 plot(mod)
 
 # Plot RGL device with constructs visible and named
 plot(mod, constructs = TRUE,
      construct.lab = c("Compassion", "Fairness", "Conformity"))
 
-# Item W7Q16 has location 6 in the data set (gender and age excluded)
+plot(mod, constructs = TRUE,
+     items = c(15, 17, 18, 19, 20),
+     construct.lab = c("Compassion", "Fairness", "Conformity"))
+
+# Item W7Q16 has location 16 in the data set (gender and age excluded)
 # Below, the item is plotted together with construct to aid the visual interpretation
 plot(mod, constructs = TRUE,
-     items = 6,
+     items = 16,
      construct.lab = c("Compassion", "Fairness", "Conformity"))
 
 # Plot RGL device on item difficulty level 5
 plot(mod, diff.level = 5)
-
-# A selection of Conformity items from the model plotted with constructs
-plot(mod, constructs = TRUE,
-     items = c(5,7,8,9,10),
-     construct.lab = c("Compassion", "Fairness", "Conformity"))
 
 # Plot RGL device with scaled items and constructs visible and named
 plot(mod, scale = TRUE,

[…]
