### 2024-06-18 #6523 @smasongarrison
> My experience with goodpractice is that it often gives false positives, and it's really hard to make it completely satisfied. Furthermore, there is no requirement in JOSS that you use the goodpractice package

@ForsbergPyschometrics

That is my experience with gp as well. Instead, I think of it as a good extra set of eyes to let you know if there's something you should look at. You may decide that it's not feasible, which is totally fine. My intent with mentioning good practice was more of a helpful tool to bring things to your attention.  It might not be feasible to refactor the plot.D3mirt function, and that's totally ok. For example, I have a function right now that has a complexity of 128; it's not feasible to slice it up because it's reading in a gedcom file. It was useful to let me know that I should think about slicing it up, but after thinking about it, I concluded that refactoring it would be harder to maintain, so I didn't.
