+++ DESCRIPTION
@@ -17,7 +17,7 @@ Suggests:
     knitr,
     rmarkdown,
     stats,
-    testthat (>= 3.2.1)
+    testthat
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.3.1
