### 2024-06-18 ForsbergPyschometrics/D3mirt#19 @ForsbergPsychometrics
Two test units have been added that cover all functions in D3mirt, version 2.0.0 (currently available on GitHub). Because the method is highly visual, the test suit relies mainly on expect_snapshot() from the test_that package.

The approach is to check functionality against the empirical data from the built-in dataset "anes0809offwaves" as well as against a new data frame included in the package containing standard angles in three dimensions reported in cartesian and spherical coordinates (data frame is called "angles"). The angles data frame is used for explicitly controlling the most essential calculations done in the D3mirt and the modid functions with the test unit. However, it can also be used to visually inspect the results with the plot function if you are familiar with how standard angles can be represented numerically as well as visually. The core of DMIRT is to transform factor loadings into angles, so using standard angles should just run straight through if the function does what it is supposed to do. This provides an external check on the package.

Other issues:
- Sapply() has been replaced with vapply() in all functions.
- The plot function has been revised and is now at complexity 85. The new version of the D3mirt() function is at complexity 92, which is due to the new features that make the workflow easier.
