### 2024-06-18 ForsbergPyschometrics/D3mirt#14 @ForsbergPsychometrics
- Test units have been added (see comment regarding issue "Functionality")
- Guidelines for contributing to code have been added to README under the heading "Getting Help, Feedback, and Questions"
- Most of the statement of need is in the README but is compressed. This will be addressed again after revising the paper.
- Examples in the documentation are now standalone.
