+++ paper.md
@@ -17,6 +17,7 @@ output:
 authors:
 - name: Erik Forsberg
   orcid: "0000-0002-5228-9729"
+  equal-contrib: yes
   affiliation: '1'
 bibliography: refs.bib
 affiliations:
@@ -27,47 +28,51 @@ affiliations:
 
 
 # Summary
-The `D3mirt` package for `R` [@R] offers functions for analyzing questionnaire items used in psychological research in a three-dimensional latent space. The application is based on descriptive multidimensional item response theory (DMIRT) [@Reckase:2009; @Reckase:1985; @Reckase+McKinley:1991], a statistical framework that incorporates vector geometry to describe item characteristics. The method is foremost visual and the latent model can be plotted as an interactive graphical device with the help of a dedicated plot function based on `RGL` 3D visualization device system for `R` [@Adler+Murdoch:2023]. Alongside the plot function, the package also includes a dedicated model identification function that helps the user identify the DMIRT model and a model estimation function for extracting the necessary vector estimates. New additions to the DMIRT framework introduced in the `D3mirt` package include the option of studying constructs (explained below) and individual scores plotted in the three-dimensional latent model.
+`D3mirt` package for `R` [@R] offers functions for analyzing questionnaire items used in psychological research in a three-dimensional latent space. The application is based on descriptive multidimensional item response theory (DMIRT) [@Reckase:2009; @Reckase:1985; @Reckase+McKinley:1991], a statistical framework that incorporates vector geometry to describe item characteristics. The method is foremost visual and the latent model can be plotted as an interactive graphical device with the help of a dedicated plot function based on `RGL` 3D visualization device system for `R` [@Adler+Murdoch:2023]. Alongside the plot function, the package also includes a dedicated model identification function that helps the user identify the DMIRT model and a model estimation function for extracting the necessary vector estimates. New additions to the DMIRT framework introduced in the `D3mirt` package include the option of studying constructs (explained below) and individual scores plotted in the three-dimensional latent model.
 
 # Statement of need
 
 Common to most item response theory (IRT) models is the assumption of *unidimensionality*, i.e., that a test or item measures simple structures [@Hambleton+Jones+Rogers:1991]. There are, however
 many occasions where this may be improper. Consider a mathematical word problem
 [@Reckase:2009; @Reckase:1985; @Reckase+McKinley:1991].
-To solve a mathematical word problem, one must often have both verbal and mathematical skills, or abilities ($\theta$) as it is often called in the literature on item response theory.
-In other words, one's result would be a function of one's ability to read, on the one hand, and one's ability to perform numerical manipulations, on the other. Accordingly, instead of a person's location on a unidimensional 
+To solve a mathematical word problem, one must often have both verbal and mathematical skills, refereed to as abilities (denoted $\theta$) in the literature on IRT.
+In other words, one's resulting answer would be a function based on one's ability to read, on the one hand, and one's ability to perform numerical manipulations, on the other. Accordingly, instead of a person's location on a unidimensional 
 latent variable, the mathematical word problem illustrates a situation where it seems more 
 reasonable to assume that a correct response is due to the respondent's location in a 
 multidimensional latent variable space. 
 
 Descriptive multidimensional item response theory (DMIRT) [@Reckase:2009; @Reckase:1985; @Reckase+McKinley:1991] has been proposed as a data reduction technique for the just mentioned situation. The method is based on using a *compensatory model*, 
[…]
