### 2024-06-18 ForsbergPyschometrics/D3mirt#17 @ForsbergPsychometrics
Guidelines has been added in the paper, package vignette and README (the latter is, however, slow to update).
