### 2024-06-18 ForsbergPyschometrics/D3mirt#13 @ForsbergPsychometrics
A brief software comparison has been added and the reference is fixed in the latest version of the paper.
