### 2024-06-18 ForsbergPyschometrics/D3mirt#15 @ForsbergPsychometrics
Issue fixed.
