### 2024-06-18 ForsbergPyschometrics/D3mirt#16 @ForsbergPsychometrics
A brief software comparison has been added to the latest revision of the paper.
