### 2024-06-18 ForsbergPyschometrics/D3mirt#19 @smasongarrison
Nice job! Your tests are a clever solution to a difficult problem!!
