### 2024-06-18 #6523 @smasongarrison
I'm satisfied with the changes. This is a useful package :)
