+++ tests/testthat/test-D3mirt.R
@@ -18,121 +18,121 @@ test_that("Test unit D3mirt and plot", {
   row <- rowSums(row^2)
   row <- matrix(row)
   for (i in seq_along(row)){
-    expect_equal(row[i,1], 1)
+    testthat::expect_equal(row[i,1], 1)
   }
   plot(y, title = "Plot Test 1.1")
   p <- rgl::scene3d()
-  expect_snapshot(p)
+  testthat::expect_snapshot(p)
   plot(y, adjust.lab = c(0.6, -0.9), font = 2, item.lab = c(1:20),
        cex = 2, title = "Plot Test 1.2", font.col = "pink", col = c("red"),
        type = "extrusion", arrow.width = 0.8, n = 22, theta = 0.3, barblen = 0.03)
   p <- rgl::scene3d()
-  expect_snapshot(p)
+  testthat::expect_snapshot(p)
   plot(y, item.names = FALSE, title = "Plot Test 2.1")
   p <- rgl::scene3d()
-  expect_snapshot(p)
+  testthat::expect_snapshot(p)
   plot(y, item.names = FALSE, adjust.lab = c(0.6, -0.9), font = 2, item.lab = c(1:20),
        cex = 2, title = "Plot Test 2.2", font.col = "pink", col = c("red"),
        type = "extrusion", arrow.width = 0.8, n = 22, theta = 0.3, barblen = 0.03)
-  expect_snapshot(p)
+  testthat::expect_snapshot(p)
   plot(y, items = c(1:10), title = "Plot Test 3.1")
   p <- rgl::scene3d()
-  expect_snapshot(p)
+  testthat::expect_snapshot(p)
   plot(y, items = c(1:10), adjust.lab = c(0.6, -0.9), font = 2, item.lab = c(1:20),
        cex = 2, title = "Plot Test 3.2", font.col = "pink", col = c("red"),
        type = "extrusion", arrow.width = 0.8, n = 22, theta = 0.3, barblen = 0.03)
   p <- rgl::scene3d()
-  expect_snapshot(p)
+  testthat::expect_snapshot(p)
   plot(y, item.names = FALSE, items = c(1:10), title = "Plot Test 4.1")
   p <- rgl::scene3d()
-  expect_snapshot(p)
+  testthat::expect_snapshot(p)
   plot(y, item.names = FALSE, items = c(1:10), adjust.lab = c(0.6, -0.9), font = 2, item.lab = c(1:20),
        cex = 2, title = "Plot Test 4.2", font.col = "pink", col = c("red"),
        type = "extrusion", arrow.width = 0.8, n = 22, theta = 0.3, barblen = 0.03)
   p <- rgl::scene3d()
-  expect_snapshot(p)
+  testthat::expect_snapshot(p)
   plot(y, items = c(1:10), item.lab = c(1:10), title = "Plot Test 5.1")
   p <- rgl::scene3d()
-  expect_snapshot(p)
+  testthat::expect_snapshot(p)
   plot(y, items = c(1:10), item.lab = c(1:10), adjust.lab = c(0.6, -0.9), font = 2,
        cex = 2, title = "Plot Test 5.2", font.col = "pink", col = c("red"),
        type = "extrusion", arrow.width = 0.8, n = 22, theta = 0.3, barblen = 0.03)
   p <- rgl::scene3d()
-  expect_snapshot(p)
+  testthat::expect_snapshot(p)
   plot(y, item.names = FALSE, items = c(1:10), item.lab = c(1:10), title = "Plot Test 6.1")
   p <- rgl::scene3d()
-  expect_snapshot(p)
+  testthat::expect_snapshot(p)
[…]
