+++ README.md
@@ -44,7 +44,7 @@ If dichotomous items are used, the analysis is based on the
 multidimensional two-parameter logistic model (M2PL) (McKinley &
 Reckase, 1983; Reckase, 1985). If polytomous items are used, the
 analysis is extended to the multidimensional two-parameter graded
-response model (MGRM; Muraki & Carlson, 1995). The method is therefore
+response model (MGRM; Muraki & Carlson, 1995). The method is, therefore,
 limited to items that fit these item models.
 
 The estimation begins by first fitting and extracting the loadings $a$
@@ -166,10 +166,12 @@ Begin by loading the item data. Note that all outputs are available as
 ready-made package files that can be loaded directly into the R session.
 
 ``` r
+# Load Package
+library(D3mirt)
 # Load data
 data("anes0809offwaves")
 x <- anes0809offwaves
-x <- x[,3:22] # Remove columns for age and gender
+x <- x[, 3:22] # Remove columns for age and gender
 ```
 
 The `modid()` can take in raw item data or a data frame with item factor
@@ -177,7 +179,7 @@ loadings. In the default mode (`efa = TRUE`), using raw data, the
 function performs an EFA using with the help of the `mirt()` function,
 from the `mirt` package (Chalmers, 2012), with three factors as the
 default (`factors = 3`). After sorting the results following the model
-identification procedure the function finishes by presenting the model
+identification procedure, the function finishes by presenting the model
 identification items. The function throws an error if no items can be
 located. Note that it is possible to use available factor loadings
 assigned to a data frame in the function call. This can be done by
@@ -195,17 +197,17 @@ generally be manipulated.
 
 ``` r
 # Optional: Load the EFA data for this example directly from the package file
-load(system.file("extdata/efa.Rdata", package = "D3mirt"))
+load(system.file("extdata/id.Rdata", package = "D3mirt"))
 ```
 
 ``` r
 # Call to modid() with x, containing factors scores from the EFA
 # Observe that the efa argument is set to false 
-a <- modid(x)
+id <- modid(x)
 ```
 
 ``` r
-summary(a)
+summary(id)
 #> 
 #> modid: 20 items and 3 factors
 #> 
@@ -335,11 +337,11 @@ load(system.file("extdata/mod3.Rdata", package = "D3mirt"))
 con <- list(c(1:10),
             c(11:14),
             c(15:20))
-g <- D3mirt(x, modid = c("W7Q3", "W7Q20"), con.items = con)
+mod3 <- D3mirt(x, modid = c("W7Q3", "W7Q20"), con.items = con)
 ```
 
 ``` r
-summary(g)
+summary(mod3)
 #> 
 #> D3mirt: 20 items and 5 levels of difficulty
 #> 
@@ -469,15 +471,15 @@ load(system.file("extdata/mod2.Rdata", package = "D3mirt"))
 ```
 
 ``` r
-y <- data.frame(x[,-16])# Remove misfitting item W7Q16
+y <- data.frame(x[, -16])# Remove misfitting item W7Q16
 # Call D3mirt() and using the orthogonal model
-h <- D3mirt(y, modid = list(c(1:10),   # x-axis
-                             c(15:19),  # y-axis
-                             c(11:14))) # z-axis
+mod2 <- D3mirt(y, modid = list(c(1:10),   # x-axis
+                            c(15:19),  # y-axis
+                            c(11:14))) # z-axis
 ```
 
 ``` r
-summary(h)
+summary(mod2)
 #> 
 #> D3mirt: 19 items and 5 levels of difficulty
 #> 
@@ -603,26 +605,28 @@ An example of how the output can be described is as follows.
 > we find Fairness ($\theta = 49.101^{\circ}$, $\phi = 78.313^{\circ}$)
[…]
