+++ vignettes/IntrotoD3mirt.Rmd
@@ -156,10 +156,8 @@ If the default mode `efa = TRUE` is not changed, the function performs an EFA wi
 
 ```{r, include = FALSE, eval = FALSE}
 load("id.Rdata")
-x <- a$loadings
-a <- modid(x, efa = FALSE)
 ```
-```{r, eval = FALSE}
+```{r, eval = TRUE}
 # Optional: Load the modid data for this example directly from the package file
 load(system.file("extdata/id.Rdata", package = "D3mirt"))
 ```
@@ -196,7 +194,7 @@ Sometimes, however, the model is hard to identify. If this happens, try the foll
 
 The latter (point 4) should, however, only be used as a last resort. This is because the upper bound sets the upper limit for item inclusion. Adjusting this limit too high compromises the necessary statistical requirements. The lower limit (point 2), however, only increases the size of the item pool used for the item selection. It is, therefore, recommended to adjust the lower limit up and down to see if the output differs, and from that make the final decision.
 
-```{r, eval = FALSE}
+```{r, eval = TRUE}
 # Assigns loadings from a to data frame x
 x <- id$loadings
 ```
@@ -213,7 +211,7 @@ Another option (point 3) is to override the factor order with the `fac.order` ar
 
 Since the squared factor loadings for factors 2 and 3 in this example are somewhat similar, it could be useful to compare the final results after manually overriding the factor order. The latter is, however, outside of the scope of this vignette.
 
-```{r, eval = FALSE}
+```{r, eval = TRUE}
 # Override factor order by reversing columns in the original data frame
 b <- modid(x, efa = FALSE, fac.order = c(3, 2, 1))
 summary(b)
@@ -247,10 +245,10 @@ To use the compensatory model, the user only needs to specify the model identifi
 
 Note that because fitting the compensatory model in the `mirt()` function takes a long time, the item parameters for this example are contained in a data frame available in the package file "mod1.Rdata". To save time studying these examples, the package file can be loaded and used in the function call. 
 
-```{r, include = FALSE}
+```{r, include = FALSE, eval = FALSE}
 load("mod1.Rdata")
 ```
-```{r, eval = FALSE}
+```{r, eval = TRUE}
 # Optional: Load mod1 directly from the package file
 load(system.file("extdata/mod1.Rdata", package = "D3mirt"))
 ```
@@ -276,10 +274,10 @@ To tell `D3mirt` to use the orthogonal model, create a nested list of items spec
 
 On a side note, using an orthogonal model in `D3mirt` often requires removing poorly fitting items. This is because constraining items already struggling to fit the model will most likely increase the misfit. This will most likely lead to a very large $MDIFF$ range, distorting the graphical output. In the example below, item W7Q16 was removed before fitting the model.
 
-```{r, include = FALSE}
+```{r, include = FALSE, eval = FALSE}
 load("mod2.Rdata")
 ```
-```{r, eval = FALSE}
+```{r, eval = TRUE}
 # Optional: Load mod2 directly from the package file
 load(system.file("extdata/mod2.Rdata", package = "D3mirt"))
 ```
@@ -306,10 +304,10 @@ This type of construct can contribute to the analysis by visualizing the average
 
[…]
