+++ DESCRIPTION
@@ -17,10 +17,11 @@ Suggests:
     knitr,
     rmarkdown,
     stats,
-    testthat
+    testthat.
+    R.rsp
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.3.1
 LazyData: true
-VignetteBuilder: knitr
+VignetteBuilder: R.rsp
 Config/testthat/edition: 3
