+++ vignettes/IntrotoD3mirt.Rmd
@@ -1,0 +1,685 @@
+---
+title: "Introduction to D3mirt Analysis"
+package: D3mirt
+output: rmarkdown::pdf_document
+vignette: >
+  %\VignetteIndexEntry{Introduction to D3mirt Analysis}
+  %\VignetteEncoding{UTF-8}
+  %\VignetteEngine{knitr::rmarkdown}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+```{r setup, message = FALSE, include = FALSE}
+library(D3mirt)
+knitr::knit_hooks$set(webgl = hook_webgl)
+```
+
+::: {align="center"}
+<img src="https://github.com/ForsbergPyschometrics/D3mirt/raw/main/man/figures/logo.png" align="center" height="300"/>
+:::
+
+# `D3MIRT` Modeling
+
+The `D3mirt` analysis is based on descriptive multidimensional item response theory (DMIRT; Reckase, 2009, 1985; Reckase & McKinley, 1991) and can be used to analyze dichotomous and polytomous items to measure latent traits (denoted $\theta$) in a three-dimensional ability space. The method is foremost visual and illustrates item characteristics with the help of vector geometry.
+
+In DMIRT analysis, it is assumed that items in a multidimensional ability space can measure single or multiple latent traits (Reckase, 2009, 1985). This assumption is realized in the *compensatory model*, i.e., a type of measurement model that uses linear combinations of $\theta$-values for ability assessment (Reckase, 2009). DMIRT builds on the results from the former model and seeks to maximize item discrimination in the latent space by relaxing the assumption of unidimensionality. The method is *descriptive* because the results describe both the strength of discrimination of the items and the extent to which the items are unidimensional, i.e., that the items discriminate on one dimension only, or are within-multidimensional, i.e., that the maximum item discrimination is reached when the item measures more than one dimension.
+
+Regarding vector orientation, the angle of the vector arrows indicates what traits, located along the axes in the model, an item measures (Reckase, 2009, 1985; Reckase & McKinley, 1991). For instance, in a two-dimensional space, an item is *unidimensional* if its item vector arrow is at $0°$ with respect to one of the axes in the model and $90°$ with respect to the other. Such an item describes a single trait only. In contrast, an item is perfectly *within-multidimensional* if its item vector arrow is oriented at $45°$ in relation to the axes in the model. Such an item describes both traits in the model equally well. The same criteria are extended to the $n$-dimensional case.
+
+The DMIRT approach uses two types of item models, dependent on item type. If dichotomous items are used, the analysis is based on the multidimensional two-parameter logistic model (M2PL) (McKinley & Reckase, 1983; Reckase, 1985). If polytomous items are used, the analysis is extended to the multidimensional two-parameter graded response model (MGRM; Muraki & Carlson, 1995). In both cases, the estimation process consists of first fitting a compensatory multidimensional two-parameter model from which loadings ($a$) and difficulty parameters ($d$) are extracted. In the case of `D3mirt`, this implies that a compensatory three-dimensional M2PL or MGRM is fitted. In the next step, the algorithm tries to find the angle and distance from the origin of the model indicating the location of the highest possible discrimination value, also referred to as the inflection point, for each item. The location, in turn, shows what latent trait, or traits, the item can be said to measure and on what level of difficulty. The output is visualized as vector arrows representing the domain of the item response functions located in a multidimensional space pointing in the direction of the maximum slope.
+
[…]
