### 2024-06-21 3de4477
Update plot.D3mirt.Rd
