+++ man/plot.D3mirt.Rd
@@ -220,7 +220,7 @@ In the the latter case, the model perspective in the still shoot can be manually
 # Load data
 data("anes0809offwaves")
 x <- anes0809offwaves
-x <- x[, 3:22] # Remove columns for age and gender
+x <- as.matrix(x[, 3:22]) # Remove columns for age and gender
 
 # Call D3mirt() with constructs assigned to con
 con <- list(c(1,2,3,4,5,6,7,8,9,10),
