### 2024-06-21 c200966
Update NEWS.md
