### 2024-06-21 c37a029
Update D3mirt.R
