### 2024-06-21 aa0712d
Delete test-D3mirt.R
