+++ man/plot.D3mirt.Rd
@@ -220,7 +220,7 @@ In the the latter case, the model perspective in the still shoot can be manually
 # Load data
 data("anes0809offwaves")
 x <- anes0809offwaves
-x <- as.matrix(x[, 3:22]) # Remove columns for age and gender
+x <- x[, 3:22] # Remove columns for age and gender
 
 # Call D3mirt() with constructs assigned to con
 con <- list(c(1,2,3,4,5,6,7,8,9,10),
@@ -260,7 +260,7 @@ plot(mod, hide = TRUE, ind.scores = TRUE)
 # Plot respondents separated on gender
 # Start by assigning the gender variable to a data frame
 # In this example, this is done by sub-setting the gender column
-x <- as.matrix(anes0809offwaves)
+x <- anes0809offwaves
 
 # Call plot() and use the gender variable column in the levels argument
 # Respondent data on gender is in column two, x[, 2]
@@ -271,8 +271,7 @@ plot(mod, hide = TRUE, ind.scores = TRUE,
     sphere.col = c("blue", "red"),
     x.lab = "Compassion",
     y.lab="Conformity",
-    z.lab="Fairness",
-    view = c(16, 20, 0.6))
+    z.lab="Fairness"))
 
 # Add a 95\% CI to respondent factor scores on <= 30 y.o.
 # Column bind trait scores with the age variable "W3Xage"
@@ -296,8 +295,7 @@ plot(mod, hide = TRUE,
     z.lab="Fairness",
     ci = TRUE,
     ci.level = 0.95,
-    ellipse.col = "orange",
-    view = c(15, 20, 0.6))
+    ellipse.col = "orange"))
 }
 \dontrun{
 # Export an open RGL device to the console to be saved as HTML or image file
