### 2024-06-22 589cbb4
Increment version number to 2.0.4
