+++ tests/testthat/test-D3mirt.R
@@ -1,0 +1,202 @@
+test_that("Test unit D3mirt and plot", {
+  data(anes0809offwaves)
+  anes1 <- as.matrix(anes0809offwaves[,3:22])
+  x <- D3mirt(anes1, modid= c("W7Q3", "W7Q20"), con.items = list(c(1:10), (11:15), c(16:20)))
+  expect_s3_class(x, "D3mirt")
+  expect_snapshot(x)
+  expect_snapshot(print(x))
+  expect_snapshot(summary(x))
+  row <- as.matrix(x$dir.cos)
+  row <- rowSums(row^2)
+  row <- matrix(row)
+  for (i in seq_along(row)){
+    expect_equal(row[i,1], 1)
+  }
+  anes2 <- data.frame(ifelse(anes1 > 4 ,1 , 0))
+  y <- D3mirt(anes2, modid= c("W7Q3", "W7Q20"), itemtype = "2PL")
+  row <- as.matrix(y$dir.cos)
+  row <- rowSums(row^2)
+  row <- matrix(row)
+  for (i in seq_along(row)){
+    testthat::expect_equal(row[i,1], 1)
+  }
+  plot(y, title = "Plot Test 1.1")
+  p <- rgl::scene3d()
+  testthat::expect_snapshot(p)
+  plot(y, adjust.lab = c(0.6, -0.9), font = 2, item.lab = c(1:20),
+       cex = 2, title = "Plot Test 1.2", font.col = "pink", col = c("red"),
+       type = "extrusion", arrow.width = 0.8, n = 22, theta = 0.3, barblen = 0.03)
+  p <- rgl::scene3d()
+  testthat::expect_snapshot(p)
+  plot(y, item.names = FALSE, title = "Plot Test 2.1")
+  p <- rgl::scene3d()
+  testthat::expect_snapshot(p)
+  plot(y, item.names = FALSE, adjust.lab = c(0.6, -0.9), font = 2, item.lab = c(1:20),
+       cex = 2, title = "Plot Test 2.2", font.col = "pink", col = c("red"),
+       type = "extrusion", arrow.width = 0.8, n = 22, theta = 0.3, barblen = 0.03)
+  testthat::expect_snapshot(p)
+  plot(y, items = c(1:10), title = "Plot Test 3.1")
+  p <- rgl::scene3d()
+  testthat::expect_snapshot(p)
+  plot(y, items = c(1:10), adjust.lab = c(0.6, -0.9), font = 2, item.lab = c(1:10),
+       cex = 2, title = "Plot Test 3.2", font.col = "pink", col = c("red"),
+       type = "extrusion", arrow.width = 0.8, n = 22, theta = 0.3, barblen = 0.03)
+  p <- rgl::scene3d()
+  testthat::expect_snapshot(p)
+  plot(y, item.names = FALSE, items = c(1:10), title = "Plot Test 4.1")
+  p <- rgl::scene3d()
+  testthat::expect_snapshot(p)
+  plot(y, item.names = FALSE, items = c(1:10), adjust.lab = c(0.6, -0.9), font = 2, item.lab = c(1:10),
+       cex = 2, title = "Plot Test 4.2", font.col = "pink", col = c("red"),
+       type = "extrusion", arrow.width = 0.8, n = 22, theta = 0.3, barblen = 0.03)
+  p <- rgl::scene3d()
+  testthat::expect_snapshot(p)
+  plot(y, items = c(1:10), item.lab = c(1:10), title = "Plot Test 5.1")
+  p <- rgl::scene3d()
+  testthat::expect_snapshot(p)
+  plot(y, items = c(1:10), item.lab = c(1:10), adjust.lab = c(0.6, -0.9), font = 2,
[…]
