### 2024-06-24 ForsbergPyschometrics/D3mirt#14 @JeffreyCHoover
This looks great! Thanks for considering these! As far as my original concerns, I think this issue is ready to be closed.
