### 2024-06-24 ForsbergPyschometrics/D3mirt#13 @JeffreyCHoover
Looks great! I think this issue can be closed down.
