### 2024-06-29 #6523 @osorensen
👋 @hendr1km, are you satisfied with the changes to this submission? If so, could you please update [your checklist](https://github.com/openjournals/joss-reviews/issues/6523#issuecomment-2093727576) to approve?
