### 2024-07-02 #6523 @osorensen
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.18637/jss.v048.i06 is OK
> - 10.1080/15366367.2023.2282418 is OK
> - 10.1177/014662169501900109 is OK
> - 10.1007/978-0-387-89976-3 is OK
>
> MISSING DOIs
>
> - No DOI given, and none found for title: plink: An R Package for Linking Mixed-Format Tests...
> - No DOI given, and none found for title: flexMIRT User’s Manual Version 3.52: Flexible Mult...
> - No DOI given, and none found for title: IRTPRO for Windows
> - No DOI given, and none found for title: Mathematica, Version 13.3
> - No DOI given, and none found for title: MATLAB Version: 9.13.0 (R2022b)
> - No DOI given, and none found for title: Mplus User’s Guide. Eight Edition
> - 10.32614/cran.package.rgl may be a valid DOI for title: Rgl: 3D Visualization Using OpenGL
> - No DOI given, and none found for title: Fundamentals of Item Response Theory
> - No DOI given, and none found for title: An Extension of the Two-parameter Logistic Model t...
> - No DOI given, and none found for title: R: A Language and Environment for Statistical Comp...
> - 10.1177/014662168500900409 may be a valid DOI for title: The Difficulty of Test Items That Measure More Tha...
> - 10.1177/014662169101500407 may be a valid DOI for title: The Discriminating Power of Items That Measure Mor...
>
> INVALID DOIs
>
> - None
> ```

@ForsbergPyschometrics, could you please go through this list and add potentially missing DOIs to your references, and notify me here when done?
