+++ refs.bib
@@ -25,7 +25,8 @@
     year = {2010},
     volume = {35},
     number = {12},
-    pages = {1--33},
+    pages = {1-33},
+    doi = {10.18637/jss.v035.i12},
     url = {http://www.jstatsoft.org/v35/i12/}}
     
 @manual{flexMIRT,
@@ -137,6 +138,7 @@ publisher = {Routledge},
 	title = {{The Difficulty of Test Items That Measure More Than One Ability}},
 	volume = {4},
 	year = {1985},
+	doi = {10.1177/014662168500900409}, 
 	bdsk-url-1 = {10.1177/014662168500900409}}
 
 @article{Reckase+McKinley:1991,
@@ -148,4 +150,5 @@ publisher = {Routledge},
 	title = {{The Discriminating Power of Items That Measure More Than One Dimension}},
 	volume = {4},
 	year = {1991},
+	doi = {10.1177/014662169101500407},
 	bdsk-url-1 = {10.1177/014662169101500407}}
