### 2024-07-02 #6523 @ForsbergPyschometrics
Hi!
I have checked all missing doi's and only three references had doi
numbers that are now included. These were the plink package by Weeks
(2010)  and the last two articles by Reckase (1985) and Reckase
McKinley (1991).

However, I get a "––––" in the reference list where it should read
Reckase, M. D. (2009) when I knit with R Studio. I have no idea why R
markdown does this. When I read the PDF draft you exported, however,
this error does not appear.

Regards,

Erik Forsberg
Psychometrician

Phone: +46(0)736007084
https://github.com/ForsbergPyschometrics

Please, consider the environment before printing this E-mail!

Regards,

Erik Forsberg
Psychometrician

Phone: +46(0)736007084
https://github.com/ForsbergPyschometrics

Please, consider the environment before printing this E-mail!

Den tis 2 juli 2024 kl 12:08 skrev The Open Journals editorial robot
***@***.***>:
>
> 👉📄 Download article proof 📄 View article proof on GitHub 📄 👈
>
