### 2024-07-02 8b90bf9
Formatting changes to paper
Editorial comment on [JOSS submission](https://github.com/openjournals/joss-reviews/issues/6523).
