### 2024-07-02 #6523 @osorensen
@ForsbergPyschometrics I'm reading through the manuscript now and will make pull requests or issues for suggested changes.
