### 2024-07-02 e71bebc
More edits
Figures will be automatically labelled, so remove the hard-coding.

https://github.com/openjournals/joss-reviews/issues/6523
