### 2024-07-02 90e444a
Equations should be part of the text flow
Not sure I covered everything in this PR, but equations are part of the sentences, also when they are displayed. 

https://github.com/openjournals/joss-reviews/issues/6523
