### 2024-07-02 46c2122
Delete paper.Rmd
