+++ paper.Rmd
@@ -1,101 +1,0 @@
----
-title: "D3mirt: Descriptive Three-Dimensional Multidimensional Item Response Theory
-  for R"
-tags:
-- R
-- descriptive multidimensional item response theory
-- test construction
-- item analysis
-- psychology
-- psychometry
-date: "10 May 2024"
-output:
-  pdf_document: default
-  html_document:
-    df_print: paged
-    keep_md: true
-authors:
-- name: Erik Forsberg
-  orcid: "0000-0002-5228-9729"
-  equal-contrib: yes
-  affiliation: '1'
-bibliography: refs.bib
-affiliations:
-- name: Division of Personality, Social and Developmental Psychology, Stockholm University,
-    Sweden
-  index: 1
----
-```{r setup, include=FALSE}
-knitr::opts_chunk$set(echo = TRUE)
-```
-
-# Summary
-The `D3mirt` package for `R` [@R] offers functions for analyzing questionnaire items used in psychological research in a three-dimensional latent space. The application is based on descriptive multidimensional item response theory (DMIRT) [@Reckase:2009; @Reckase:1985; @Reckase+McKinley:1991], a statistical framework incorporating vector geometry to describe item characteristics. The method is foremost visual, and the latent model can be plotted as an interactive graphical device with the help of a dedicated plot function based on the `RGL` 3D visualization device system for `R` [@Adler+Murdoch:2023]. Alongside the plot function, the package also includes a dedicated model identification function that helps the user identify the DMIRT model and a model estimation function for extracting the necessary vector estimates. New additions to the DMIRT framework introduced in the `D3mirt` package include studying constructs (explained below) and individual scores plotted in the three-dimensional latent model.
-
-# Statement of need
-
-Common to most item response theory (IRT) models is the assumption of *unidimensionality*, i.e., that a test or item measures simple structures [@Hambleton+Jones+Rogers:1991]. There are, however,
-many occasions where this may be improper. Consider a mathematical word problem
-[@Reckase:2009; @Reckase:1985; @Reckase+McKinley:1991].
-To solve a mathematical word problem, one must often have verbal and mathematical skills, referred to as abilities (denoted $\theta$) in the literature on IRT.
-In other words, one's resulting answer would be a function based on one's ability to read, on the one hand, and one's ability to perform numerical manipulations, on the other. Accordingly, instead of a person's location on a unidimensional 
-latent variable, the mathematical word problem illustrates a situation where it seems more 
-reasonable to assume that a correct response is due to the respondent's location in a 
-multidimensional latent variable space. 
-
-Descriptive multidimensional item response theory (DMIRT) [@Reckase:2009; @Reckase:1985; @Reckase+McKinley:1991] has been proposed as a data reduction technique for the just mentioned situation. The method is based on using a *compensatory model*, 
-i.e., a type of measurement model within multidimensional IRT that uses linear combinations of $\theta$-values for ability assessment.
-Accordingly, this model assumes that the same probability score for a correct response can be reached by adding different combinations of $\theta$-values. In turn, this implies that the compensatory model allows items to be unidimensional, i.e., that they measure a single ability, or *within-multidimensional*, i.e., that the items measure more than one ability in the model space.
-
-Note that the `D3mirt` approach is limited to two types of compensatory models, depending on item type.
-If dichotomous items are used, the analysis is based on the multidimensional extension of the two-parameter logistic model [@McKinley+Reckase:1983] as the compensatory model. If polytomous items are used, the analysis is based on the two-parameter multidimensional graded response model [@Muraki+Carlson:1995] as the compensatory model.
-
[…]
