+++ paper.Rmd
@@ -54,40 +54,40 @@ Compared to other software, `D3mirt` is unique in implementing DMIRT methodology
 
 The `D3mirt` package was designed to counter many of the just mentioned shortages by implementing specialized functions for identifying the DMIRT model, calculation of the necessary DMIRT estimates, and plotting the results in an interactive three-dimensional latent environment (see \autoref{fig:anes}). An example of the utility of using the package in an empirical context for item and scale analysis has been presented in @Forsberg:2024.
 
-![**Figure 1**. A still shot of the graphical output from `D3mirt`. The Figure illustrates a three-dimensional vector plot for items in the `anes0809offwaves` data set included in the package. The output also shows three construct vector arrows: Compassion, Fairness, and Conformity (solid black arrows).\label{fig:anes}](anes1.png){ width=100% }
+![A still shot of the graphical output from `D3mirt`. The Figure illustrates a three-dimensional vector plot for items in the `anes0809offwaves` data set included in the package. The output also shows three construct vector arrows: Compassion, Fairness, and Conformity (solid black arrows).\label{fig:anes}](anes1.png){ width=100% }
 
-![**Figure 2**. A still shot of the graphical output from `D3mirt` illustrating respondent scores in the latent space separated on sex (male in blue and female in red) from the `anes0809offwaves` data set included in the package.\label{fig:p1}](p1.png){ width=100% }
+![A still shot of the graphical output from `D3mirt` illustrating respondent scores in the latent space separated on sex (male in blue and female in red) from the `anes0809offwaves` data set included in the package.\label{fig:p1}](p1.png){ width=100% }
 
 # Multidimensional item parameters
 
 The theoretical framework for DMIRT rests foremost on three assumptions [@Reckase:1985]. Firstly, ability maps the probability monotonically, such that a higher level of ability implies a higher probability of answering an item correctly. Second, we wish to locate an item at a singular point at which it is possible to derive item characteristics for the multidimensional case. Thirdly, an item's maximum level of discrimination, i.e., its highest possible sensitivity score for measuring ability, is the best option for the singular point estimation. The most important parameter equations capturing the assumptions mentioned regarding DMIRT will be presented below.
 
-Firstly, by using the discrimination score $a_i$ from the compensatory model, we can derive the multidimensional analog to the unidimensional discrimination parameter, i.e., the multidimensional discrimination index (MDISC), denoted $A_i$ for item $i$, to highlight the connection to the unidimensional $a_i$ parameter [@Reckase:2009; @Reckase+McKinley:1991].
+Firstly, by using the discrimination score $a_i$ from the compensatory model, we can define the multidimensional analog to the unidimensional discrimination parameter, i.e., the multidimensional discrimination index (MDISC), denoted $A_i$ for item $i$, to highlight the connection to the unidimensional $a_i$ parameter [@Reckase:2009; @Reckase+McKinley:1991].
 
 \begin{equation} \label{eq:MDISC}
-MDISC := \sqrt {\sum_{k = 1}^{m} a^{2}_{ik}}
+MDISC := \sqrt {\sum_{k = 1}^{m} a^{2}_{ik}},
 \end{equation}
 
-With the slope constant $\frac{1}{4}$ omitted [@Reckase:2009; @Reckase+McKinley:1991]. Importantly, the MDISC sets the orientation of the item vectors in the multidimensional space [@Reckase:2009; @Reckase+McKinley:1991] as follows.
[…]
