+++ paper.md
@@ -76,7 +76,7 @@ with the slope constant $\frac{1}{4}$ omitted [@Reckase:2009; @Reckase+McKinley:
 
 on latent axis $l$ in the model. Note, the $\omega_{il}$ is in this solution a characteristic of the item $i$ that tells in what direction $i$ has its highest level of discrimination, assuming a multidimensional latent space [@Reckase:2009; @Reckase+McKinley:1991]. This gives us the following criteria to use as a rule of thumb. Assume a two-dimensional space, an orientation of $0^{\circ}$ with respect to any of the model axes indicates that the item is unidimensional. Such an item describes a singular trait only. In contrast, an orientation of $45^{\circ}$ indicated that the item is within-multidimensional. Such an item describes both traits in the two-dimensional model equally well. The same criteria are extended to the three-dimensional case. The MDISC is also used in the graphical output to scale the length of the vector arrows representing the item response functions, e.g., so that longer vector arrows indicate higher discrimination, shorter arrows lower discrimination in the model, and so on.
 
-Next, to assess multidimensional difficulty, the distance from the origin is calculated using the multidimensional difficulty (MDIFF), denoted $B_i$, index [@Reckase:1985]:
+Next, to assess multidimensional difficulty, the distance from the origin is calculated using the multidimensional difficulty (MDIFF), denoted $B_i$, index [@Reckase:1985].
 
 \begin{equation} \label{eq:MDIFF}
 MDIFF :=\frac{-d_i}{\sqrt{\sum_{k=1}^m a^2_{ik}}},
