+++ README.Rmd
@@ -35,7 +35,7 @@ The $MDIFF$ is interpreted similarly to the difficulty parameter in the unidimen
 
 The $MDISC$ shows the highest level of discrimination an item can achieve in the multidimensional latent model. It is, therefore, a global item characteristic referring only to the discrimination achieved in the specific direction of the item in the latent space. The $MDISC$ score is visualized in the graphical output by scaling the length of the vector arrows (longer arrows indicate higher discrimination and vice versa), which in turn describe the domain of item response functions.
 
-In `D3mirt`it also possible to use constructs in the analysis. Constructs, in this context, refer to the assumption that a subset of items or a particular orientation in the model space can measure a higher-order latent variable. This is implemented in `D3mirt` as optional vectors whose orientation is either calculated as the average direction of a subset of items (from all items down to a single item) or indicated by spherical coordinates. A construct vector will, therefore, point in the direction of the maximum slope of what we might think of as an imaginary item response function indicated by the items or coordinates chosen by the user based on exploratory or theoretical reasons. 
+In `D3mirt`it also possible to use constructs in the analysis. Constructs, in this context, refer to the assumption that a subset of items or a particular orientation in the model space can measure a higher-order latent variable. This is implemented in `D3mirt` as optional vectors whose orientation is either calculated as the average direction of a subset of items (from all items down to a single item) or indicated by spherical coordinates. A construct vector will, therefore, point in the direction of the maximum slope of what we might think of as an imaginary item response function indicated by the items or coordinates chosen by the user based on exploratory or theoretical reasons.
 
 If constructs are used, the output will include reporting the directional discrimination ($DDISC$) parameter that shows how well the items discriminate, assuming that they measure one of the constructs used in the analysis. That is, while the $MDISC$ represents the maximum level of discrimination in the model, the $DDISC$ represents the local discrimination that makes it possible to compare item discrimination in a specific direction set by the constructs. The constructs are, therefore, like unidimensional models nested in the multidimensional latent space and are visually represented with construct vector arrows scaled to an arbitrary length.
 
@@ -90,7 +90,7 @@ x <- anes0809offwaves
 x <- x[, 3:22] # Remove columns for age and gender
 ```
 
-The `modid()` can take in raw item data or a data frame with item factor loadings. In the default mode (`efa = TRUE`), using raw data, the function performs an EFA using with the help of the `mirt()` function, from the `mirt` package (Chalmers, 2012), with three factors as the default (`factors = 3`). After sorting the results following the model identification procedure, the function finishes by presenting the model identification items. The function throws an error if no items can be located. Note that it is possible to use available factor loadings assigned to a data frame in the function call. This can be done by setting `efa ` to `FALSE`, which allows the function to skip the EFA and jump directly to the model identification procedure.
[…]
