+++ README.md
@@ -970,3 +970,5 @@ Reckase, M. D., & McKinley, R. L. (1991). The Discriminating Power of
 Items That Measure More Than One Dimension. *Applied Psychological
 Measurement, 15*(4), 361-373.
 <https://doi.org/10.1177/014662169101500407>
+
+[![](https://github.com/ForsbergPyschometrics/D3mirt/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/ForsbergPyschometrics/D3mirt/actions/workflows/R-CMD-check.yaml)
