+++ .github/workflows/paper.yml
@@ -1,4 +1,5 @@
 on: [push]
+on: workflow_dispatch
 
 jobs:
   paper:
