### 2024-07-02 0668e9f
Update paper.yml
