+++ .github/workflows/paper.yml
@@ -1,5 +1,4 @@
 on: [push]
-on: workflow_dispatch
 
 jobs:
   paper:
