+++ paper.md
@@ -84,7 +84,7 @@ MDIFF :=\frac{-d_i}{\sqrt{\sum_{k=1}^m a^2_{ik}}},
 
 in which $d$ is the $d$-parameter index from the compensatory model. The MDIFF is denoted $B$ as the DMIRT counterpart to the $b$-parameter in the unidimensional IRT model. The MDIFF is, therefore, a characteristic of item $i$ such that higher MDIFF values indicate that higher levels of ability are necessary for a correct response  [@Reckase:2009;  @Reckase+McKinley:1991]. Observe that the denominator in \autoref{eq:MDIFF} is the same expression as \autoref{eq:MDISC}. 
 
-Importantly, in DMIRT analysis, the MDISC and MDIFF only apply in the direction set by $\omega_{il}$ and Equation \autoref{eq:dcos} [@Reckase:2009; @Reckase+McKinley:1991]. Thus, we cannot compare these estimates directly across items, as would be the case in the unidimensional model. This is because DMIRT seeks to maximize item discrimination as a global characteristic in a multidimensional environment. To estimate item discrimination as a local characteristic in the multidimensional space, it is, however, possible to select a common direction for the items and then recalculate the discrimination, i.e., to estimate the directional discrimination (DDISC), as follows:
+Importantly, in DMIRT analysis, the MDISC and MDIFF only apply in the direction set by $\omega_{il}$ and \autoref{eq:dcos} [@Reckase:2009; @Reckase+McKinley:1991]. Thus, we cannot compare these estimates directly across items, as would be the case in the unidimensional model. This is because DMIRT seeks to maximize item discrimination as a global characteristic in a multidimensional environment. To estimate item discrimination as a local characteristic in the multidimensional space, it is, however, possible to select a common direction for the items and then recalculate the discrimination, i.e., to estimate the directional discrimination (DDISC), as follows:
 
 \begin{equation} \label{eq:DDISC}
 DDISC :=\sum_{k=1}^{m}a_{ik}cos\,\omega_{ik}.
@@ -92,7 +92,9 @@ DDISC :=\sum_{k=1}^{m}a_{ik}cos\,\omega_{ik}.
 
 Since the DDISC is a local characteristic in the model, it is always the case that $DDISC\,\leq\,MDISC$. In `D3mirt`, the DDISC is optional and implemented in `D3mirt` as optional *construct vectors* indicated by a subset of items or using spherical coordinates. 
 
-The results include tables for the MDISC and MDIFF estimates as well as spherical coordinates describing the location of the vector arrows. If construct vectors are used, the output also includes DDISC scores for all items showing the constrained discrimination. It is also possible to plot individual scores (i.e., *profile analysis*) in the three-dimensional latent space (see \autoref{fig:p1}). This can be useful for studying respondents' location conditioned on some external variable, e.g., sex, age, political preference, and so on. Instructions on the method, such as model identification, model estimation, plotting, and profile analysis, are given in the package vignette. To report issues, seek support, or for developers wishing to contribute to the software, contact the author via the dedicated GitHub page (<https://github.com/ForsbergPyschometrics/D3mirt>) or email ([forsbergpsychometrics\@gmail.com](mailto:forsbergpsychometrics@gmail.com){.email}). 
[…]
