+++ .github/workflows/R-CMD-check.yaml
@@ -49,4 +49,4 @@ jobs:
       - uses: r-lib/actions/check-r-package@v2
         with:
           upload-snapshots: true
-          build_args: 'c("--no-manual","--compact-vignettes=gs+qpdf","--no-build-vignettes")'
+          build_args: 'c('--no-manual','--compact-vignettes=gs+qpdf','--no-build-vignettes')'
