### 2024-07-03 f4cdce4
Update paper.md
