+++ paper.md
@@ -25,12 +25,9 @@ affiliations:
     Sweden
   index: 1
 ---
-```{r setup, include=FALSE}
-knitr::opts_chunk$set(echo = TRUE)
-```
 
 # Summary
-The `D3mirt` package for `R` [@R] offers functions for analyzing questionnaire items used in psychological research in a three-dimensional latent space. The application is based on descriptive multidimensional item response theory (DMIRT) [@Reckase:2009; @Reckase:1985; @Reckase+McKinley:1991], a statistical framework incorporating vector geometry to describe item characteristics. The method is foremost visual, and the latent model can be plotted as an interactive graphical device with the help of a dedicated plot function based on the `RGL` 3D visualization device system for `R` [@Adler+Murdoch:2023]. Alongside the plot function, the package also includes a dedicated model identification function that helps the user identify the DMIRT model and a model estimation function for extracting the necessary vector estimates. New additions to the DMIRT framework introduced in the `D3mirt` package include studying constructs (explained below) and individual scores plotted in the three-dimensional latent model.
+The `D3mirt` package for `R` [@R] offers functions for analyzing questionnaire items used in psychological research in a three-dimensional latent space. The application is based on descriptive multidimensional item response theory (DMIRT) [@Reckase:2009; @Reckase:1985; @Reckase+McKinley:1991], a statistical framework incorporating vector geometry to describe item characteristics. The method is foremost visual, and the latent model can be plotted as an interactive graphical device with the help of a dedicated plot function based on the `RGL` 3D visualization device system for `R` [@Adler+Murdoch:2023]. Alongside the plot function, the package also includes a model identification function that helps the user identify the DMIRT model and a model estimation function for extracting the necessary vector estimates. New additions to the DMIRT framework introduced in the `D3mirt` package include studying constructs (explained below) and individual scores plotted in the three-dimensional latent model.
 
 # Statement of need
 
@@ -43,9 +40,8 @@ latent variable, the mathematical word problem illustrates a situation where it 
 reasonable to assume that a correct response is due to the respondent's location in a 
 multidimensional latent variable space. 
 
-Descriptive multidimensional item response theory (DMIRT) [@Reckase:2009; @Reckase:1985; @Reckase+McKinley:1991] has been proposed as a data reduction technique for the just mentioned situation. The method is based on using a *compensatory model*, 
-i.e., a type of measurement model within multidimensional IRT that uses linear combinations of $\theta$-values for ability assessment.
-Accordingly, this model assumes that the same probability score for a correct response can be reached by adding different combinations of $\theta$-values. In turn, this implies that the compensatory model allows items to be unidimensional, i.e., that they measure a single ability, or *within-multidimensional*, i.e., that the items measure more than one ability in the model space.
+Descriptive multidimensional item response theory (DMIRT) [@Reckase:2009; @Reckase:1985; @Reckase+McKinley:1991] was developed to handle the just mentioned situation. The method is based on using a *compensatory model*, 
[…]
