+++ .github/workflows/R-CMD-check.yaml
@@ -49,4 +49,6 @@ jobs:
       - uses: r-lib/actions/check-r-package@v2
         with:
           upload-snapshots: true
-          build_args: 'c("--no-manual","--compact-vignettes=gs+qpdf","--no-build-vignettes")'
+          build_args: 'c("--no-manual", "--no-build-vignettes", "--as-cran")'
+          error-on: '"error"'
+          check-dir: '"check"'
