+++ .github/workflows/R-CMD-check.yaml
@@ -44,7 +44,7 @@ jobs:
 
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
-          extra-packages: any::rcmdcheck
+          extra-packages: '("any::rcmdcheck", "rgl")'
           needs: check
 
       - uses: r-lib/actions/check-r-package@v2
