+++ .github/workflows/R-CMD-check.yaml
@@ -11,6 +11,7 @@ name: R-CMD-check
 permissions: read-all
 
 jobs:
+       
   R-CMD-check:
     runs-on: ${{ matrix.config.os }}
 
@@ -31,6 +32,8 @@ jobs:
       R_KEEP_PKG_SOURCE: yes
 
     steps:
+      - if: runner.os == 'macOS'
+        run: brew install --cask xquartz
       
       - uses: actions/checkout@v4
 
@@ -44,7 +47,7 @@ jobs:
 
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
-          extra-packages: '("any::rcmdcheck", "rgl")'
+          extra-packages: any::rcmdcheck rgl
           needs: check
 
       - uses: r-lib/actions/check-r-package@v2
@@ -53,6 +56,3 @@ jobs:
           build_args: 'c("--no-manual", "--no-build-vignettes")'
           error-on: '"error"'
           check-dir: '"check"'
-
-      - if: runner.os == 'macOS'
-        run: brew install --cask xquartz
