+++ DESCRIPTION
@@ -21,7 +21,7 @@ Suggests:
     R.rsp
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
-RoxygenNote: 7.3.1
+RoxygenNote: 7.3.2
 LazyData: true
 VignetteBuilder: R.rsp
 Config/testthat/edition: 3
