### 2024-07-04 #6523 @ForsbergPyschometrics
@osorensen, I must apologize for the slight confusion. I've been changing computers, and an older version of the manuscript must have been accidentally pushed to Github. Nevertheless, the current version should contain all the revisions asked for. I also read through and changed a few passages slightly to improve the readabilty of the paper been further. All automated workflow run check out in the main branch and I've also added the R-CMD and the JOSS badges to the README file.
