### 2024-07-04 #6523 @ForsbergPyschometrics
- Tagged release GitHub: v2.0.4
- DOI for archived version: DOI 10.17605/OSF.IO/ZEFGW
