### 2024-07-04 #6523 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
