+++ README.Rmd
@@ -19,9 +19,8 @@ library(D3mirt)
 <img src="man/figures/logo.png" align="center" height="300"/>
 :::
 
-::: {align="left"}
+\
 [![](https://github.com/ForsbergPyschometrics/D3mirt/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/ForsbergPyschometrics/D3mirt/actions/workflows/R-CMD-check.yaml) [![status](https://joss.theoj.org/papers/abc1dfe0f4805bdc11f1e01b0937957b/status.svg)](https://joss.theoj.org/papers/abc1dfe0f4805bdc11f1e01b0937957b)
-:::
 
 # `D3MIRT` Modeling
 
