### 2024-07-15 #6523 @samhforbes
Hi @ForsbergPyschometrics I can see editorialbot has flagged a possible DOI for RGL (it seems to resolve to the CRAN page, but I can't see it in the citation info in R). Was this DOI something you were looking to include (I can't see exactly where it comes from, but you might know more)?
