### 2024-07-15 #6523 @ForsbergPyschometrics
Hi @samhforbes, I was unaware that the RGL package had a DOI but I noticed now on the CRAN page. Should I include it?
