+++ refs.bib
@@ -84,6 +84,7 @@ publisher = {Routledge},
 	note = {[Computer Software]},
 	title = {Rgl: 3D Visualization Using OpenGL},
 	url = {https://dmurdoch.github.io/rgl/index.html},
+	doi = {10.32614/CRAN.package.rgl},
 	bdsk-url-1 = {https://dmurdoch.github.io/rgl/index.html}}
 	
 @book{Hambleton+Jones+Rogers:1991,
