### 2024-07-18 #6523 @samhforbes
Hi @ForsbergPyschometrics I think if you edit the version you have until you are happy, and try to generate pdf here to check, we can then potentially re-issue the accept and it should refresh I think. Do you want to give this a try? We have to use the version created by editorialbot
