### 2024-07-18 #6523 @ForsbergPyschometrics
@samhforbes, the current version of the article uploadad on my Github is the accepted article with the new edits. So assume a re-issue and accept is only what is required? If you check the pdf you will see the changes.
