### 2024-07-22 #6523 @ForsbergPyschometrics
Hi @samhforbes, That was an improvement but sadly it introduced a very unfortunate page break. I have reuploaded another version in which the image size is slightly reduced. Is it possible to try again? Hopefully, this will be the last.
