### 2024-08-01 #6523 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
