### 2024-08-02 #6523 @ForsbergPyschometrics
Hi,
Thank you for your help with the article. Just a general question, can I
remove the files connected to the review process from my GitHub? I was
thinking of the draft papers in different formats and figures.

Regards,

Erik Forsberg
Psychometrician

Phone: +46(0)736007084
https://github.com/ForsbergPyschometrics

Please, consider the environment before printing this E-mail!

Den tors 1 aug. 2024 kl 13:23 skrev The Open Journals editorial robot <
***@***.***>:

> 🎉🎉🎉 Congratulations on your paper acceptance! 🎉🎉🎉
>
> If you would like to include a link to your paper from your README use the
> following code snippets:
>
> Markdown:
>
> HTML:
> <a style="border-width:0" href="https://doi.org/10.21105/joss.06523">
> </a>
>
> reStructuredText:
>    :target: https://doi.org/10.21105/joss.06523
>
> This is how it will look in your documentation:
>
> [image: DOI] <https://doi.org/10.21105/joss.06523>
>
> *We need your help!*
>
> The Journal of Open Source Software is a community-run journal and relies
> upon volunteer effort. If you'd like to support us please consider doing
> either one (or both) of the the following:
>
>    - Volunteering to review for us sometime in the future. You can add
>    your name to the reviewer list here:
>    https://reviewers.joss.theoj.org/join
>    - Making a small donation to support our running costs here:
>    https://numfocus.org/donate-to-joss
>
